% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{addCls}
\alias{addCls}
\alias{addCls,AnalysisData-method}
\alias{addCls,Analysis-method}
\title{addCls}
\usage{
addCls(x, cls, value, ...)

\S4method{addCls}{AnalysisData}(x, cls, value)

\S4method{addCls}{Analysis}(x, cls, value, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{name of new sample information column}

\item{value}{new sample information to add}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Add a sample info column to a Analysis or AnalysisData object.
}
