% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateMean}
\alias{aggregateMean}
\alias{aggregateMean,AnalysisData-method}
\alias{aggregateMedian}
\alias{aggregateMedian,AnalysisData-method}
\alias{aggregateSum}
\alias{aggregateSum,AnalysisData-method}
\title{Sample aggregation}
\usage{
aggregateMean(d, cls = "class")

\S4method{aggregateMean}{AnalysisData}(d, cls = "class")

aggregateMedian(d, cls = "class")

\S4method{aggregateMedian}{AnalysisData}(d, cls = "class")

aggregateSum(d, cls = "class")

\S4method{aggregateSum}{AnalysisData}(d, cls = "class")
}
\arguments{
\item{d}{S4 object of class \code{AnalysisData}}

\item{cls}{info column to use for class data}
}
\value{
An S4 object of class \code{AnalysisData} containing the transformed data.
}
\description{
Aggregation of sample features based on a grouping variable.
}
\details{
Sample aggregation allows the electronic pooling of sample features based on a grouping variable.
This is useful in situations such as the presence of technical replicates that can be aggregated to reduce the effects of pseudo replication.
}
\section{Methods}{

\itemize{
\item \code{aggregateMean}: Aggregate sample features to the group mean.
\item \code{aggregateMedian}: Aggregate sample features to the group median.
\item \code{aggregateSum}: Aggregate sample features to the group total.
}
}

\examples{
## Each of the following examples shows the application of the aggregation method and then 
## a Principle Component Analysis is plotted to show it's effect on the data structure.

## Initial example data preparation
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\% 
 occupancyMaximum(occupancy = 2/3)

d \%>\% 
 plotPCA(cls = 'day')
 
## Mean aggregation
d \%>\% 
 aggregateMean(cls = 'day') \%>\% 
 plotPCA(cls = 'day',ellipses = FALSE)
 
## Median aggregation
d \%>\% 
 aggregateMedian(cls = 'day') \%>\% 
 plotPCA(cls = 'day',ellipses = FALSE)
 
## Sum aggregation
d \%>\% 
 aggregateSum(cls = 'day') \%>\% 
 plotPCA(cls = 'day',ellipses = FALSE)
 
}
