% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/aggregateMethods.R
\docType{methods}
\name{aggregateMean}
\alias{aggregateMean}
\alias{aggregateMean,AnalysisData-method}
\title{aggregateMean}
\usage{
aggregateMean(d, cls = "class")

\S4method{aggregateMean}{AnalysisData}(d, cls = "class")
}
\arguments{
\item{d}{S4 object of class Data}

\item{cls}{info column to use for class data}
}
\description{
Mean aggregation of sample data.
}
