% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-accessors.R
\name{dat}
\alias{dat}
\alias{dat,AnalysisData-method}
\alias{dat,Analysis-method}
\alias{dat<-}
\alias{dat<-,AnalysisData-method}
\alias{dat<-,Analysis-method}
\alias{sinfo}
\alias{sinfo,AnalysisData-method}
\alias{sinfo,Analysis-method}
\alias{sinfo<-}
\alias{sinfo<-,AnalysisData-method}
\alias{sinfo<-,Analysis-method}
\alias{raw}
\alias{raw,Analysis-method}
\alias{raw<-}
\alias{raw<-,Analysis-method}
\alias{preTreated}
\alias{preTreated,Analysis-method}
\alias{preTreated<-}
\alias{preTreated<-,Analysis-method}
\alias{features}
\alias{features,AnalysisData-method}
\alias{features,Analysis-method}
\alias{nSamples}
\alias{nSamples,AnalysisData-method}
\alias{nSamples,Analysis-method}
\alias{nFeatures}
\alias{nFeatures,AnalysisData-method}
\alias{nFeatures,Analysis-method}
\alias{analysisResults}
\alias{analysisResults,Analysis-method}
\title{\code{AnalysisData} and \code{Analysis} class accessors}
\usage{
dat(x, ...)

\S4method{dat}{AnalysisData}(x)

\S4method{dat}{Analysis}(x, type = c("pre-treated", "raw"))

dat(x, ...) <- value

\S4method{dat}{AnalysisData}(x) <- value

\S4method{dat}{Analysis}(x, type = c("pre-treated", "raw")) <- value

sinfo(x, ...)

\S4method{sinfo}{AnalysisData}(x)

\S4method{sinfo}{Analysis}(x, type = c("pre-treated", "raw"), value)

sinfo(x, ...) <- value

\S4method{sinfo}{AnalysisData}(x) <- value

\S4method{sinfo}{Analysis}(x, type = c("pre-treated", "raw")) <- value

raw(x)

\S4method{raw}{Analysis}(x)

raw(x) <- value

\S4method{raw}{Analysis}(x) <- value

preTreated(x)

\S4method{preTreated}{Analysis}(x)

preTreated(x) <- value

\S4method{preTreated}{Analysis}(x) <- value

features(x, ...)

\S4method{features}{AnalysisData}(x)

\S4method{features}{Analysis}(x, type = c("pre-treated", "raw"))

nSamples(x, ...)

\S4method{nSamples}{AnalysisData}(x)

\S4method{nSamples}{Analysis}(x, type = c("pre-treated", "raw"))

nFeatures(x, ...)

\S4method{nFeatures}{AnalysisData}(x)

\S4method{nFeatures}{Analysis}(x, type = c("pre-treated", "raw"))

analysisResults(x, element)

\S4method{analysisResults}{Analysis}(x, element)
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData} or \code{Analysis}}

\item{...}{arguments to pass to the appropriate method}

\item{type}{get or set \code{raw} or \code{pre-treated} data}

\item{value}{value to set}

\item{element}{analysis element results to return}
}
\description{
Accessor methods for the \code{AnalysisData} and \code{Analysis} S4 classes.
}
\section{Methods}{

\itemize{
\item \code{dat}: Return a metabolomic data table.
\item \verb{dat<-}: Set a metabolomic data table.
\item \code{sinfo}: Return a sample information data table.
\item \verb{sinfo<-}: Set a sample information data table.
\item \code{raw}: Return the \code{AnalysisData} object containing unprocessed metabolomic data from an \code{Analysis} object.
\item \verb{raw<-}: Set an \code{AnalysisData} object to the \code{raw} slot of an \code{Analysis} class object.
\item \code{preTreated}: Return the \code{AnalysisData} object containing pre-treated metabolomic data from an \code{Analysis} object.
\item \verb{preTreated<-}: Set an \code{AnalysisData} object to the \code{pre-treated} slot of an \code{Analysis} class object.
\item \code{features}: Return the features names.
\item \code{nSamples}: Return the number of samples.
\item \code{nFeatures}: Return the number of features.
\item \code{analysisResults}: Return results from an \code{Analysis} object of an analysis element.
}
}

\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Return the metabolomic data
dat(d)

## Set the metabolomic data
dat(d) <- abr1$neg[,300:400]

## Return the sample information
sinfo(d)

## Set the sample information
sinfo(d) <- abr1$fact

## Return the feature names
features(d)

## Return the number of samples
nSamples(d)

## Return the number of features
nFeatures(d)
}
