% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AnalysisData-class}
\alias{AnalysisData-class}
\title{AnalysisData S4 class}
\description{
An S4 class for metabolomic data and sample meta information.
}
\section{Slots}{

\describe{
\item{\code{data}}{sample metabolomic data}

\item{\code{info}}{sample meta information}
}}

