% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisData.R
\name{analysisData}
\alias{analysisData}
\title{AnalysisData class constructor}
\usage{
analysisData(data, info)
}
\arguments{
\item{data}{table containing sample metabolomic data}

\item{info}{table containing sample meta information}
}
\value{
An S4 object of class Analysis.
}
\description{
Create an AnalysisData S4 object.
}
\examples{
library(metaboData)
d <- analysisData(data = abr1$neg,info = abr1$fact)

print(d)
}
