% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{analysisElements}
\alias{analysisElements}
\title{Analysis elements}
\usage{
analysisElements()
}
\value{
A character vector of analysis elements.
}
\description{
Return the analysis elements available in \code{metabolyseR}.
}
\examples{
analysisElements()
}
