% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisParameters.R
\name{analysisParameters}
\alias{analysisParameters}
\title{analysisParameters}
\usage{
analysisParameters(elements = c("preTreat", "classification",
  "featureSelection", "correlations"))
}
\description{
analysisParameters
}
