% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AnalysisPlot-class}
\alias{AnalysisPlot-class}
\title{AnalysisPlot}
\description{
An S4 class to store an analysis plot
}
\section{Slots}{

\describe{
\item{\code{data}}{list containing data used to generate the plot}

\item{\code{plot}}{ggplot object containing the plot}

\item{\code{func}}{function used to generate the plot}
}}

