% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisResults.R
\name{analysisResults}
\alias{analysisResults}
\alias{analysisResults,Analysis-method}
\title{analysisResults}
\usage{
analysisResults(x, element)

\S4method{analysisResults}{Analysis}(x, element)
}
\arguments{
\item{x}{S4 object of class Analysis}

\item{element}{Analysis element to extract. 
Should be one of those returned \code{analysisElements()}.}
}
\description{
Extract analysis results for a given analysis element.
}
