% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{anova}
\alias{anova}
\alias{anova,AnalysisData-method}
\title{ANOVA}
\usage{
anova(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE
)

\S4method{anova}{AnalysisData}(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE
)
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{a vector of sample info column names to analyse}

\item{pAdjust}{p value adjustment method}

\item{comparisons}{list of comparisons to perform}

\item{returnModels}{should models be returned}
}
\description{
One-way analysis of variance (ANOVA).
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Perform ANOVA
anova_analysis <- anova(d,cls = 'day')

## Extract significant features
explanatoryFeatures(anova_analysis)
}
