% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{availableCls}
\alias{availableCls}
\alias{availableCls,AnalysisData-method}
\alias{availableCls,Analysis-method}
\title{availableCls}
\usage{
availableCls(x, ...)

\S4method{availableCls}{AnalysisData}(x)

\S4method{availableCls}{Analysis}(x, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Return available sample info columns from an Analysis or AnalysisData object.
}
