% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\name{binaryComparisons}
\alias{binaryComparisons}
\alias{binaryComparisons,AnalysisData-method}
\title{binaryComparisons}
\usage{
binaryComparisons(x, cls = "class")

\S4method{binaryComparisons}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class AnalysisData.}

\item{cls}{sample information column to use}
}
\description{
Return a vector of possible binary comparisons for a 
given sample information column.
}
