% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{bindRows}
\alias{bindRows}
\alias{bindRows,list-method}
\title{Bind \code{AnalysisData} objects by row}
\usage{
bindRows(d)

\S4method{bindRows}{list}(d)
}
\arguments{
\item{d}{list object containing S4 objects of class AnalysisData to be bound}
}
\value{
An S4 object of class AnalysisData containg the bound data sets.
}
\description{
Bind the rows of AnalysisData objects contained within a list.
}
\examples{
library(metaboData)
d <- list(
 negative = analysisData(abr1$neg,abr1$fact),
 positive = analysisData(abr1$pos,abr1$fact)
 )

bindRows(d)
}
