% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/join.R
\name{bindAnalysesRows}
\alias{bindAnalysesRows}
\alias{bindAnalysesRows,list-method}
\title{bindAnalysesRows}
\usage{
bindAnalysesRows(x)

\S4method{bindAnalysesRows}{list}(x)
}
\arguments{
\item{x}{list object containing S4 objects of class AnalysisData to be bound}
}
\description{
Bind rows of objects of class AnalysisData contained within a list.
}
