% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{changeParameter<-}
\alias{changeParameter<-}
\alias{changeParameter<-,AnalysisParameters-method}
\title{Change analysis parameters}
\usage{
changeParameter(x, parameterName, elements = analysisElements()) <- value

\S4method{changeParameter}{AnalysisParameters}(x, parameterName, elements = analysisElements()) <- value
}
\arguments{
\item{x}{S4 object of class \code{AnalysisParameters}}

\item{parameterName}{name of the parameter to change}

\item{elements}{character vector of analysis elements to target parameter
change. Can be any returned by \code{analysisElements()}.}

\item{value}{New value of the parameter}
}
\value{
An S4 object of class \code{AnalysisParameters}.
}
\description{
Change analysis parameters.
}
\details{
For the parameter name selected, all parameters with that name will
be altered.
}
\examples{
p <- analysisParameters('pre-treatment')

changeParameter(p,'cls') <- 'day'

print(p)
}
