% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classification-method.R
\docType{methods}
\name{classification}
\alias{classification}
\alias{classification,Analysis-method}
\title{classification}
\usage{
classification(x)

\S4method{classification}{Analysis}(x)
}
\description{
classification
}
