% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classificationResults-method.R
\docType{methods}
\name{classificationResults}
\alias{classificationResults}
\alias{classificationResults,Analysis-method}
\title{classificationResults-Analysis}
\usage{
classificationResults(x)

\S4method{classificationResults}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract classification results from an Analysis object
}
