% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{clsAdd}
\alias{clsAdd}
\alias{clsAdd,AnalysisData-method}
\alias{clsAdd,Analysis-method}
\title{clsAdd}
\usage{
clsAdd(x, cls, value, ...)

\S4method{clsAdd}{AnalysisData}(x, cls, value)

\S4method{clsAdd}{Analysis}(x, cls, value, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{name of new sample information column}

\item{value}{new sample information to add}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Add a sample info column to a Analysis or 
AnalysisData object.
}
