% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{clsArrange}
\alias{clsArrange}
\alias{clsArrange,AnalysisData-method}
\alias{clsArrange,Analysis-method}
\title{clsArrange}
\usage{
clsArrange(x, cls = "class", descending = FALSE, ...)

\S4method{clsArrange}{AnalysisData}(x, cls = "class", descending = FALSE)

\S4method{clsArrange}{Analysis}(x, cls = "class", descending = FALSE, type = "raw")
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{cls}{name of sample information column to arrange by}

\item{descending}{TRUE/FALSE, arrange samples in descending order}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Order samples within an object of class AnalysisData or Analysis by a given sample information column.
}
