% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{clsAvailable}
\alias{clsAvailable}
\alias{clsAvailable,AnalysisData-method}
\alias{clsAvailable,Analysis-method}
\title{clsAvailable}
\usage{
clsAvailable(x, ...)

\S4method{clsAvailable}{AnalysisData}(x)

\S4method{clsAvailable}{Analysis}(x, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Return available sample info columns from an Analysis or
 AnalysisData object.
}
