% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{clsRemove}
\alias{clsRemove}
\alias{clsRemove,AnalysisData-method}
\alias{clsRemove,Analysis-method}
\title{clsRemove}
\usage{
clsRemove(x, cls, ...)

\S4method{clsRemove}{AnalysisData}(x, cls)

\S4method{clsRemove}{Analysis}(x, cls, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{name of sample information column to remove}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Remove a sample info column from a Analysis or 
AnalysisData object.
}
