% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{clsRename}
\alias{clsRename}
\alias{clsRename,AnalysisData-method}
\alias{clsRename,Analysis-method}
\title{clsRename}
\usage{
clsRename(x, cls, newName, ...)

\S4method{clsRename}{AnalysisData}(x, cls, newName)

\S4method{clsRename}{Analysis}(x, cls, newName, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{sample information column to rename}

\item{newName}{new column name}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Rename a sample information column within an object of AnalysisData or Analysis.
}
