% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{correctionCenter}
\alias{correctionCenter}
\alias{correctionCenter,AnalysisData-method}
\title{Batch/block correction}
\usage{
correctionCenter(d, block = "block", type = c("mean", "median"))

\S4method{correctionCenter}{AnalysisData}(d, block = "block", type = c("mean", "median"))
}
\arguments{
\item{d}{S4 object of class \code{AnalysisData}}

\item{block}{sample information column name to use containing sample block
groupings}

\item{type}{type of average to use}
}
\value{
An S4 object of class \code{AnalysisData} containing the corrected data.
}
\description{
Correction of batch/block differences.
}
\details{
There can sometimes be artificial batch related variability introduced into metabolomics analyses as a result of analytical instrumentation or sample preparation.
With an appropriate randomised block design of sample injection order, batch related variability can be corrected using an average centring correction method of the individual features.
}
\section{Methods}{

\itemize{
\item \code{correctionCenter}: Correction using group average centring.
}
}

\examples{

## Initial example data preparation
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\% 
 occupancyMaximum(occupancy = 2/3)
 
## Group total ion count distributions prior to correction
d \%>\% 
 plotTIC(by = 'day',colour = 'day')
 
## Group total ion count distributions after group median correction
d \%>\% 
 correctionCenter(block = 'day',type = 'median') \%>\% 
 plotTIC(by = 'day',colour = 'day')
}
