% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/correctionMethods.R
\name{correctionCenter}
\alias{correctionCenter}
\alias{correctionCenter,AnalysisData-method}
\title{correctionCenter}
\usage{
correctionCenter(d, block = "block", type = "median")

\S4method{correctionCenter}{AnalysisData}(d, block = "block", type = "median")
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{block}{info column containing sample block 
groupings to use for correction}

\item{type}{averaging to use; eg. mean or median}
}
\description{
Batch correction using average centering.
}
