% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/correlationResults-method.R
\docType{methods}
\name{correlationResults}
\alias{correlationResults}
\alias{correlationResults,Analysis-method}
\title{correlationResults-Analysis}
\usage{
correlationResults(x)

\S4method{correlationResults}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract correlation results from an Analysis object
}
