% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/correlations-method.R
\docType{methods}
\name{correlations}
\alias{correlations}
\alias{correlations,Analysis-method}
\title{correlations}
\usage{
correlations(x)

\S4method{correlations}{Analysis}(x)
}
\description{
correlations
}
