% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{correlationsParameters}
\alias{correlationsParameters}
\title{Correlations parameters}
\usage{
correlationsParameters()
}
\description{
Retrieve the default parameters for correlation analysis.
}
\examples{
## Retrieve the default correlation parameters
p <- correlationsParameters()

## Assign the correlation parameters to analysis parameters
cp <- analysisParameters('correlations')
parameters(cp,'correlations') <- p

print(cp)
}
