% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\docType{methods}
\name{dat}
\alias{dat}
\alias{dat,AnalysisData-method}
\title{dat}
\usage{
dat(x)

\S4method{dat}{AnalysisData}(x)
}
\arguments{
\item{x}{S4 object of class Data}
}
\description{
Return sample data from an AnalysisData object.
}
