% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\docType{methods}
\name{data}
\alias{data}
\alias{data,Data-method}
\title{data}
\usage{
data(x)

\S4method{data}{Data}(x)
}
\arguments{
\item{x}{S4 object of class Data}
}
\description{
Return sample data from a Data object.
}
