% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\docType{methods}
\name{explanatoryFeatures}
\alias{explanatoryFeatures}
\alias{explanatoryFeatures,Univariate-method}
\title{explanatoryFeatures}
\usage{
explanatoryFeatures(x, ...)

\S4method{explanatoryFeatures}{Univariate}(x, threshold)
}
\arguments{
\item{x}{S4 object of class Analysis, RandomForest or Univariate}

\item{...}{arguments to parse to method for specific class}
}
\description{
Extract explanatory features for modelling results
}
