% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters.R
\name{exportParameters}
\alias{exportParameters}
\alias{exportParameters,AnalysisParameters-method}
\alias{exportParameters,Analysis-method}
\title{exportParameters}
\usage{
exportParameters(x, file = "analysis_parameters.yaml")

\S4method{exportParameters}{AnalysisParameters}(x, file = "analysis_parameters.yaml")

\S4method{exportParameters}{Analysis}(x, file = "analysis_parameters.yaml")
}
\arguments{
\item{x}{S4 object of class AnalysisParameters or Analysis}

\item{file}{File name and path to export to}
}
\description{
Export analysis parameters from AnalysisParameters or 
Analysis objects to YAML format.
}
