% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{extractCls}
\alias{extractCls}
\alias{extractCls,AnalysisData-method}
\alias{extractCls,Analysis-method}
\title{extractCls}
\usage{
extractCls(x, cls = "class", ...)

\S4method{extractCls}{AnalysisData}(x, cls = "class")

\S4method{extractCls}{Analysis}(x, cls = "class", type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{sample info column to extract}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Extract a given sample info column from an Analysis or AnalysisData object.
}
