% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\name{features}
\alias{features}
\alias{features,AnalysisData-method}
\alias{features,Analysis-method}
\title{features}
\usage{
features(x, ...)

\S4method{features}{AnalysisData}(x)

\S4method{features}{Analysis}(x, type = "raw")
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{...}{arguments to pass to the appropriate method}

\item{type}{return features from "raw" or "pre-treated" data}
}
\description{
Return a vector of the feature names.
}
