% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/featureSelection-method.R
\docType{methods}
\name{featureSelection}
\alias{featureSelection}
\alias{featureSelection,Analysis-method}
\title{featureSelection}
\usage{
featureSelection(x)

\S4method{featureSelection}{Analysis}(x)
}
\description{
featureSelection
}
