% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/featureSelectionResults-method.R
\docType{methods}
\name{featureSelectionResults}
\alias{featureSelectionResults}
\alias{featureSelectionResults,Analysis-method}
\title{featureSelectionResults-Analysis}
\usage{
featureSelectionResults(x)

\S4method{featureSelectionResults}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract feature selection results from an Analysis object
}
