% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\alias{flatten,AnalysisData-method}
\title{Flatten an \code{AnalysisData} object}
\usage{
flatten(x)

\S4method{flatten}{AnalysisData}(x)
}
\arguments{
\item{x}{An object of S4 class \code{AnalysisData}}
}
\description{
Flatten an \code{AnalysisData} S4 class object into a single tibble.
}
\examples{
d <- analysisData(
  metaboData::abr1$neg,
  metaboData::abr1$fact
)

flatten(d)
}
