% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\docType{methods}
\name{importance,RandomForest-method}
\alias{importance,RandomForest-method}
\title{importance}
\usage{
\S4method{importance}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
return feature importance resutls from a RandomForest object
}
