% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/randomForest.R
\name{importance}
\alias{importance}
\alias{importance,RandomForest-method}
\alias{importance,Univariate-method}
\alias{importance,list-method}
\title{importance}
\usage{
importance(x)

\S4method{importance}{RandomForest}(x)

\S4method{importance}{Univariate}(x)

\S4method{importance}{list}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest or Univariate}
}
\description{
return feature importance resutls from a RandomForest or Univariate classes
}
