% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\name{importanceMetrics}
\alias{importanceMetrics}
\alias{importanceMetrics,RandomForest-method}
\title{importanceMetrics}
\usage{
importanceMetrics(x)

\S4method{importanceMetrics}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
Return available importance measures from an object 
of class RandomForest.
}
