% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputeAll}
\alias{imputeAll}
\alias{imputeAll,AnalysisData-method}
\alias{imputeClass}
\alias{imputeClass,AnalysisData-method}
\title{Missing data imputation}
\usage{
imputeAll(d, occupancy = 2/3, parallel = "variables", seed = 1234)

\S4method{imputeAll}{AnalysisData}(d, occupancy = 2/3, parallel = "variables", seed = 1234)

imputeClass(d, cls = "class", occupancy = 2/3, seed = 1234)

\S4method{imputeClass}{AnalysisData}(d, cls = "class", occupancy = 2/3, seed = 1234)
}
\arguments{
\item{d}{S4 object of class \code{AnalysisData}}

\item{occupancy}{occupancy threshold above which missing values of a feature will be imputed}

\item{parallel}{parallel type to use. See \code{?missForest} for details}

\item{seed}{random number seed}

\item{cls}{info column to use for class labels}
}
\value{
An S4 object of class \code{AnalysisData} containing the data after imputation.
}
\description{
Impute missing values using random forest imputation.
}
\details{
Missing values can have an important influence on downstream analyses with zero values heavily influencing the outcomes of parametric tests.
Where and how they are imputed are important considerations and is highly related to variable occupancy.
The methods provided here allow both these aspects to be taken into account and utilise random forest imputation using the \code{missForest} package.
}
\section{Methods}{

\itemize{
\item \code{imputeAll}: Impute missing values across all sample features.
\item \code{imputeClass}: Impute missing values class-wise.
}
}

\examples{
## Each of the following examples shows the application of each imputation method and then 
## a Linear Discriminant Analysis is plotted to show it's effect on the data structure.

## Initial example data preparation
library(metaboData)

d <- analysisData(abr1$neg[,200:250],abr1$fact) \%>\% 
 occupancyMaximum(occupancy = 2/3)

d \%>\% 
 plotLDA(cls = 'day')
 
## Missing value imputation across all samples
d \%>\% 
 imputeAll(parallel = 'no') \%>\% 
 plotLDA(cls = 'day')

## Missing value imputation class-wise
d \%>\% 
 imputeClass(cls = 'day') \%>\% 
 plotLDA(cls = 'day')
}
