% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/imputeMethods.R
\name{imputeClass}
\alias{imputeClass}
\alias{imputeClass,AnalysisData-method}
\title{imputeClass}
\usage{
imputeClass(d, cls = "class", occupancy = 2/3, seed = 1234)

\S4method{imputeClass}{AnalysisData}(d, cls = "class", occupancy = 2/3, seed = 1234)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class labels}

\item{occupancy}{occupancy threshold for imputation}

\item{seed}{random number seed}
}
\description{
Impute missing values class-wise using Random Forest.
}
