% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\docType{methods}
\name{info}
\alias{info}
\alias{info,AnalysisData-method}
\alias{info<-}
\title{info}
\usage{
info(x)

\S4method{info}{AnalysisData}(x)

info(x) <- value
}
\arguments{
\item{x}{S4 object of class Data}
}
\description{
Return sample info from an AnalysisData object.
}
