% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parseParameters}
\alias{parseParameters}
\alias{exportParameters}
\alias{exportParameters,AnalysisParameters-method}
\alias{exportParameters,Analysis-method}
\title{Parse/export analysis parameters}
\usage{
parseParameters(path)

exportParameters(d, file = "analysis_parameters.yaml")

\S4method{exportParameters}{AnalysisParameters}(d, file = "analysis_parameters.yaml")

\S4method{exportParameters}{Analysis}(d, file = "analysis_parameters.yaml")
}
\arguments{
\item{path}{file path of .yaml file to parse}

\item{d}{S4 object of class AnalysisParameters or Analysis}

\item{file}{File name and path to export to}
}
\description{
Import analysis parameters from a \code{.yaml} format file or export an \code{AnalysisParameters} object to \code{.yaml} format.
}
\examples{
## Import analysis parameters
paramFile <- system.file('defaultParameters.yaml',package = 'metabolyseR')
p <- parseParameters(paramFile)
p

\dontrun{
## Export analysis parameters
exportParameters(p,file = 'analysis_parameters.yaml')
}
}
