% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/keepMethods.R
\name{keepClasses}
\alias{keepClasses}
\alias{keepClasses,AnalysisData-method}
\title{keepClasses}
\usage{
keepClasses(d, cls = "class", classes = c())

\S4method{keepClasses}{AnalysisData}(d, cls = "class", classes = c())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class information}

\item{classes}{classes to keep}
}
\description{
Keep classes from an AnalysisData object.
}
