% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/keepMethods.R
\name{keepFeatures}
\alias{keepFeatures}
\alias{keepFeatures,AnalysisData-method}
\title{keepFeatures}
\usage{
keepFeatures(d, features = character())

\S4method{keepFeatures}{AnalysisData}(d, features = character())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{features}{features to keep}
}
\description{
Keep features from an AnalysisData object.
}
