% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/keepMethods.R
\docType{methods}
\name{keepSamples}
\alias{keepSamples}
\alias{keepSamples,AnalysisData-method}
\title{keepSamples}
\usage{
keepSamples(d, idx = "fileOrder", samples = c())

\S4method{keepSamples}{AnalysisData}(d, idx = "fileOrder",
  samples = c())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{idx}{info column containing sample indexes}

\item{samples}{sample indexes to keep}
}
\description{
Keep samples from an AnalysisData object.
}
