% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/keepMethods.R
\name{keepVariables}
\alias{keepVariables}
\alias{keepVariables,AnalysisData-method}
\title{keepVariables}
\usage{
keepVariables(d, variables = character())

\S4method{keepVariables}{AnalysisData}(d, variables = character())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{variables}{variables to keep}
}
\description{
Keep variables from an AnalysisData object.
}
