% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{linearRegression}
\alias{linearRegression}
\alias{linearRegression,AnalysisData-method}
\title{Linear regression}
\usage{
linearRegression(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  returnModels = FALSE
)

\S4method{linearRegression}{AnalysisData}(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  returnModels = FALSE
)
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{vector of sample information column names to regress}

\item{pAdjust}{p value adjustment method}

\item{returnModels}{should models be returned}
}
\value{
An S4 object of class \code{Univariate}.
}
\description{
Linear regression
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Perform linear regression
lr_analysis <- linearRegression(d,cls = 'injorder')

## Extract significant features
explanatoryFeatures(lr_analysis)
}
