% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mds}
\alias{mds}
\alias{mds,RandomForest-method}
\alias{mds,list-method}
\alias{mds,Analysis-method}
\title{Multidimensional scaling (MDS)}
\usage{
mds(x, dimensions = 2, idx = NULL)

\S4method{mds}{RandomForest}(x, dimensions = 2, idx = NULL)

\S4method{mds}{list}(x, dimensions = 2, idx = NULL)

\S4method{mds}{Analysis}(x, dimensions = 2, idx = NULL)
}
\arguments{
\item{x}{S4 object of class \code{RandomForest}, \code{Analysis} or a list}

\item{dimensions}{The number of dimensions by which the data are to be represented.}

\item{idx}{sample information column to use for sample names. If \code{NULL}, the sample row number will be used. Sample names should be unique for each row of data.}
}
\value{
A tibble containing the scaled dimensions.
}
\description{
Multidimensional scaling of random forest proximities.
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day')

mds(rf)
}
