% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\docType{methods}
\name{measures,RandomForest-method}
\alias{measures,RandomForest-method}
\title{measures}
\usage{
\S4method{measures}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
return measures results from a RandomForest object
}
