% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/randomForest.R
\name{metrics}
\alias{metrics}
\alias{metrics,RandomForest-method}
\alias{metrics,list-method}
\title{metrics}
\usage{
metrics(x)

\S4method{metrics}{RandomForest}(x)

\S4method{metrics}{list}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
return metrics results from a RandomForest object
}
