% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{modellingMethods}
\alias{modellingMethods}
\alias{modellingParameters}
\title{Modelling parameters}
\usage{
modellingMethods()

modellingParameters(methods)
}
\arguments{
\item{methods}{character vector of available modelling methods}
}
\description{
Retrieve the available modelling methods and parameters.
}
\examples{
## Retrieve the available modelling methods
modellingMethods()

## Retrieve the modelling parameters for the anova method
p <- modellingParameters('anova')

## Assign the modelling parameters to analysis parameters
mp <- analysisParameters('modelling')

parameters(mp,'modelling') <- p

print(mp)
}
