% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{modellingParameters}
\alias{modellingParameters}
\title{modellingParameters}
\usage{
modellingParameters(methods)
}
\arguments{
\item{methods}{character vector of available methods. Set to NULL to print available methods.}
}
\description{
Return parameters for a given modelling method.
}
