% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\name{modellingResults}
\alias{modellingResults}
\alias{modellingResults,Analysis-method}
\title{modellingResults}
\usage{
modellingResults(x)

\S4method{modellingResults}{Analysis}(x)
}
\arguments{
\item{x}{S4 object of class Analysis}
}
\description{
Return modelling results from an Analysis object.
}
