% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\name{nFeatures}
\alias{nFeatures}
\alias{nFeatures,AnalysisData-method}
\alias{nFeatures,Analysis-method}
\title{nFeatures}
\usage{
nFeatures(x, ...)

\S4method{nFeatures}{AnalysisData}(x)

\S4method{nFeatures}{Analysis}(x, type = "raw")
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{...}{arguments to pass to the appropriate method}

\item{type}{return features from "raw" or "pre-treated" data}
}
\description{
Return the number of features.
}
