% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\name{nSamples}
\alias{nSamples}
\alias{nSamples,AnalysisData-method}
\alias{nSamples,Analysis-method}
\title{nSamples}
\usage{
nSamples(x, ...)

\S4method{nSamples}{AnalysisData}(x)

\S4method{nSamples}{Analysis}(x, type = "raw")
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{...}{arguments to pass to the appropriate method}

\item{type}{return features from "raw" or "pre-treated" data}
}
\description{
Return the number of samoles.
}
