% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/occupancyMethods.R
\docType{methods}
\name{occupancyMaximum}
\alias{occupancyMaximum}
\alias{occupancyMaximum,AnalysisData-method}
\title{occupancyMaximum}
\usage{
occupancyMaximum(dat, cls = "class", occupancy = 2/3)

\S4method{occupancyMaximum}{AnalysisData}(dat, cls = "class",
  occupancy = 2/3)
}
\arguments{
\item{dat}{S4 object of class Data}

\item{cls}{info column to use for class data}

\item{occupancy}{occupancy threshold}
}
\description{
Maximum occupancy filtering of sample data.
}
