% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/occupancyMethods.R
\docType{methods}
\name{occupancyMinimum}
\alias{occupancyMinimum}
\alias{occupancyMinimum,AnalysisData-method}
\title{occupancyMinimum}
\usage{
occupancyMinimum(dat, cls = "class", occupancy = 2/3)

\S4method{occupancyMinimum}{AnalysisData}(dat, cls = "class",
  occupancy = 2/3)
}
\arguments{
\item{dat}{S4 object of class Data}

\item{cls}{info column to use for class data}

\item{occupancy}{occupancy threshold}
}
\description{
Minimum occupancy filtering of sample data.
}
