% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseParameters.R
\name{parseParameters}
\alias{parseParameters}
\title{parseParameters}
\usage{
parseParameters(path)
}
\arguments{
\item{path}{file path of .yaml file to parse}
}
\description{
parse .yaml file containing analysis parameters.
}
\examples{

paramFile <- system.file('defaultParameters.yaml',package = 'metabolyseR')
p <- parseParameters(paramFile)
p

}
