% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotClassification-method.R
\docType{methods}
\name{plotClassification}
\alias{plotClassification}
\alias{plotClassification,Analysis-method}
\title{plotClassification}
\usage{
plotClassification(analysis, method = "randomForest")

\S4method{plotClassification}{Analysis}(analysis, method = "randomForest")
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{method}{results of classifier to plot}
}
\description{
Plot classification results.
}
\examples{
\dontrun{
library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat','classification'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p) 
plotClassification(analysis)
}
}
