% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/plotClassificationDendrogram-method.R
\docType{methods}
\name{plotClassificationDendrogram}
\alias{plotClassificationDendrogram}
\alias{plotClassificationDendrogram,Analysis-method}
\title{plotClassificationDendrogram}
\usage{
plotClassificationDendrogram(analysis, method = "randomForest",
  measure = "Margin", clusterMethod = "ward.D2")

\S4method{plotClassificationDendrogram}{Analysis}(analysis,
  method = "randomForest", measure = "Margin", clusterMethod = "ward.D2")
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{method}{results of classifier to plot}

\item{measure}{model measure to use}

\item{clusterMethod}{clustering method to use}
}
\description{
plot a dendrogram of classifcation results
}
\examples{
\dontrun{

library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat','classification'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p) 
plotClassificationDendrogram(analysis)
}
}
