% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/plotExplanatoryHeatmap-method.R
\docType{methods}
\name{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap,Analysis-method}
\title{plotExplanatoryHeatmap}
\usage{
plotExplanatoryHeatmap(analysis, method = "fs.rf", threshold = 0.01,
  pairwises = NULL, distanceMeasure = "euclidean",
  clusterMethod = "ward.D2", colour = ptol_pal()(1))

\S4method{plotExplanatoryHeatmap}{Analysis}(analysis, method = "fs.rf",
  threshold = 0.01, pairwises = NULL, distanceMeasure = "euclidean",
  clusterMethod = "ward.D2", colour = (ggthemes::ptol_pal())(1))
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{method}{results of feature selection method to use}

\item{threshold}{score threshold to use for specifying explantory features}

\item{pairwises}{optional vector specifying pairwise comparisons to extract}

\item{distanceMeasure}{distance measure to use for clustering. See details.}

\item{clusterMethod}{clustering method to use. See details}

\item{colour}{heatmap colour to use}
}
\description{
plot a heatmap of explanatory features
}
\details{
Options for distance measures are as for \code{dist()}.
Clustering methods are as given for \code{hclust()}.
}
\examples{
\dontrun{

library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat','featureSelection'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p) 
plotExplanatoryHeatmap(analysis)
}
}
\seealso{
\link{dist} \link{hclust}
}
