% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotFeatureSelection-method.R
\docType{methods}
\name{plotFeatureSelection}
\alias{plotFeatureSelection}
\alias{plotFeatureSelection,Analysis-method}
\title{plotFeatureSelection}
\usage{
plotFeatureSelection(analysis, method = "fs.rf", mz = T, modes = T)

\S4method{plotFeatureSelection}{Analysis}(analysis, method = "fs.rf",
  mz = T, modes = T)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{method}{results of feature selection method to use}

\item{mz}{\code{TRUE} if features are m/z}

\item{modes}{split modes if present}
}
\description{
Plot feature selection results
}
\examples{
\dontrun{

library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat','featureSelection'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p) 
plotFeatureSelection(analysis)
}
}
