% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling-plots.R
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,RandomForest-method}
\alias{plotMDS,list-method}
\title{Multidimensional scaling (MDS) plot}
\usage{
plotMDS(
  x,
  cls = "class",
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotMDS}{RandomForest}(
  x,
  cls = "class",
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotMDS}{list}(
  x,
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{x}{S4 object of class \code{RandomForest}}

\item{cls}{sample information column to use for sample labelling,
Set to NULL for no labelling.}

\item{label}{sample information column to use for sample labels. Set to NULL for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points.
Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\%
confidence ellipses for each class}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument
of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot multidimensional scaling plot for a \code{RandomForest} class object.
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day')

plotMDS(rf,cls = 'day')
}
