% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingPlots.R
\name{plotMetrics}
\alias{plotMetrics}
\alias{plotMetrics,RandomForest-method}
\alias{plotMetrics,list-method}
\title{Plot model performance metrics}
\usage{
plotMetrics(x, response = "class")

\S4method{plotMetrics}{RandomForest}(x)

\S4method{plotMetrics}{list}(x)
}
\arguments{
\item{x}{S4 object of class \code{RandomForest}}

\item{response}{response results to plot}
}
\description{
Plot random forest model performance metrics
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       keepClasses(cls = 'day',classes = c('H','1','5')) \%>\% 
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day',binary = TRUE)

plotMetrics(rf,response = 'day')
}
