% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotOccupancy.R
\name{plotOccupancy}
\alias{plotOccupancy}
\alias{plotOccupancy,AnalysisData-method}
\alias{plotOccupancy,Analysis-method}
\title{plotOccupancy}
\usage{
plotOccupancy(x, cls = "class", ...)

\S4method{plotOccupancy}{AnalysisData}(x, cls = "class")

\S4method{plotOccupancy}{Analysis}(x, cls = "class", type = "raw")
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{cls}{info column to use for class labels}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{preTreated} data to plot}
}
\description{
Plot occupancy distributions.
}
