% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\docType{methods}
\name{plotResults}
\alias{plotResults}
\alias{plotResults,Univariate-method}
\title{plotResults}
\usage{
plotResults(x, predictor = "class")

\S4method{plotResults}{Univariate}(x, predictor = "class")
}
\arguments{
\item{x}{S4 object of class Univariate}

\item{predictor}{Predictor results to plot}
}
\description{
Plot univariate modelling results.
}
