% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling-plots.R
\name{plotROC}
\alias{plotROC}
\alias{plotROC,RandomForest-method}
\alias{plotROC,list-method}
\title{Plot receiver operator characteristic (ROC) curves}
\usage{
plotROC(x, title = "", legendPosition = "bottom")

\S4method{plotROC}{RandomForest}(x, title = "", legendPosition = "bottom")

\S4method{plotROC}{list}(x, title = "", legendPosition = "bottom")
}
\arguments{
\item{x}{S4 object of class \code{RandomForest}}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position
argument of \code{ggplot2::theme}. Set to "none" to remove legend.}
}
\description{
Plot receiver operator characteristic curves for a
\code{RandomForest} class object.
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day')

plotROC(rf)
}
