% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotRSD-method.R
\name{plotRSD}
\alias{plotRSD}
\alias{plotRSD,AnalysisData-method}
\alias{plotRSD,Analysis-method}
\title{plotRSD}
\usage{
plotRSD(analysis, cls = "class", ...)

\S4method{plotRSD}{AnalysisData}(analysis, cls = "class")

\S4method{plotRSD}{Analysis}(
  analysis,
  cls = "class",
  QCidx = "QC",
  QCparameters = NULL,
  histBins = 30,
  title = ""
)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for class labels}

\item{...}{arguments to pass to the appropriate method}

\item{QCidx}{QC sample label}

\item{QCparameters}{alternative parameters for QC sample pre-treatment. See details}

\item{histBins}{number of bins to use for histogram plotting}

\item{title}{plot title}
}
\description{
Plot RSD distributions of raw data in quality control samples.
}
\details{
If QCparameters is set as \code{NULL}, the default QC pre-treatment parameters are used as given by \code{analysisParameters('preTreat')}. Alternative pre-treatment routines can be used by specifying an \code{AnalysisParameters} object for \code{QCparameters}.
}
\examples{
\dontrun{

library(metaboData)
library(binneR)

files <- filePaths("FIE-HRMS","BdistachyonEcotypes")

info <- runinfo("FIE-HRMS","BdistachyonEcotypes")

binDat <- binneRlyse(files, 
                       info, 
                       parameters = detectParameters(files))

p <- analysisParameters('preTreat')

analysis <- metabolyse(binnedData(binDat)$n,info(binDat),p)

plotRSD(analysis)
}
}
