% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotTIC-method.R
\docType{methods}
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,Analysis-method}
\title{plotTIC}
\usage{
plotTIC(analysis, by = "injOrder", colour = "batchBlock", modes = T)

\S4method{plotTIC}{Analysis}(analysis, by = "injOrder",
  colour = "batchBlock", modes = T)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{by}{info column to plot against}

\item{colour}{info column to provide colour labels}

\item{modes}{split modes if present}
}
\description{
Plot total ion counts of sample raw data.
}
\examples{

library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotTIC(analysis,by = 'injorder',colour = 'day',modes = F) 
}
