% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{preTreatmentElements}
\alias{preTreatmentElements}
\alias{preTreatmentMethods}
\alias{preTreatmentParameters}
\title{Pre-treatment parameters}
\usage{
preTreatmentElements()

preTreatmentMethods(element)

preTreatmentParameters(methods)
}
\arguments{
\item{element}{pre-treatment element name}

\item{methods}{a named list of element methods}
}
\description{
Return pre-treatment elements, methods and parameters.
}
\examples{
## Return the availalble pre-treatment elements
preTreatmentElements()

## Return the available pre-treatment methods for the remove element
preTreatmentMethods('remove')

## Define some default pre-treatment parameters
p <- preTreatmentParameters(
  list(
    remove = 'classes',
    QC = c('RSDfilter','removeQC'),
    transform = 'TICnorm'
  )
)

## Assign the pre-treatment parameters to analysis paraemeters
ap <- analysisParameters('pre-treatment')
parameters(ap,'pre-treatment') <- p

print(ap)
}
