% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/preTreatedData-method.R
\docType{methods}
\name{preTreatedData}
\alias{preTreatedData}
\alias{preTreatedData,Analysis-method}
\title{preTreatedData-Analysis}
\usage{
preTreatedData(x)

\S4method{preTreatedData}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract pre-treated data from an Analysis object
}
