% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTreatedInfo-method.R
\docType{methods}
\name{preTreatedInfo,Analysis-method}
\alias{preTreatedInfo,Analysis-method}
\title{preTreatedInfo-Analysis}
\usage{
\S4method{preTreatedInfo}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract pre-treated info from an Analysis object
}
