% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{preTreatmentParameters}
\alias{preTreatmentParameters}
\title{preTreatmentParameters}
\usage{
preTreatmentParameters(methods)
}
\arguments{
\item{methods}{a named list of element methods}
}
\description{
Return default parameters for given pre-treatment element methods.
}
\examples{
p <- preTreatmentParameters(
  list(
    remove = 'classes',
    QC = c('RSDfilter','removeQC'),
    transform = 'TICnorm'
  )
)
}
