% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/QCMethods.R
\docType{methods}
\name{QCimpute}
\alias{QCimpute}
\alias{QCimpute,AnalysisData-method}
\title{QCimpute}
\usage{
QCimpute(d, cls = "class", QCidx = "QC", occupancy = 2/3,
  parallel = "variables", nCores = detectCores(),
  clusterType = "PSOCK", seed = 1234)

\S4method{QCimpute}{AnalysisData}(d, cls = "class", QCidx = "QC",
  occupancy = 2/3, parallel = "variables", nCores = detectCores(),
  clusterType = "PSOCK", seed = 1234)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{occupancy}{occupancy threshold for imputation}

\item{parallel}{parallel type to use. See `?missForest` for details}

\item{nCores}{number of cores for parallisation}

\item{clusterType}{cluster type for parallisation}

\item{seed}{random number seed}
}
\description{
QC imputation.
}
