% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/QCMethods.R
\docType{methods}
\name{QCoccupancy}
\alias{QCoccupancy}
\alias{QCoccupancy,AnalysisData-method}
\title{QCoccupancy}
\usage{
QCoccupancy(dat, cls = "class", QCidx = "QC", occupancy = 2/3)

\S4method{QCoccupancy}{AnalysisData}(dat, cls = "class", QCidx = "QC",
  occupancy = 2/3)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{occupancy}{occupancy threshold for filtering}
}
\description{
QC maximum occupancy filter.
}
