% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/QCMethods.R
\docType{methods}
\name{QCremove}
\alias{QCremove}
\alias{QCremove,AnalysisData-method}
\title{QCremove}
\usage{
QCremove(dat, cls = "class", QCidx = "QC")

\S4method{QCremove}{AnalysisData}(dat, cls = "class", QCidx = "QC")
}
\arguments{
\item{dat}{S4 object of class AnalysisData}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}
}
\description{
Remove QC samples.
}
