% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{RandomForest-class}
\alias{RandomForest-class}
\title{RandomForest S4 class}
\description{
An S4 class for random forest results and models.
}
\section{Slots}{

\describe{
\item{\code{type}}{random forest type}

\item{\code{response}}{response variable name}

\item{\code{metrics}}{tibble of model performance metrics}

\item{\code{predictions}}{tibble of model observation predictions}

\item{\code{permutations}}{list of permutations measure and importance results tables}

\item{\code{importances}}{tibble of model feature importances}

\item{\code{proximities}}{tibble of model observation proximities}

\item{\code{models}}{list of random forest models}
}}

