% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/multivariateMethods.R
\docType{methods}
\name{randomForest}
\alias{randomForest}
\alias{randomForest,AnalysisData-method}
\title{randomForest}
\usage{
randomForest(x, cls, reps = 1, returnModels = F, seed = 1234, ...)

\S4method{randomForest}{AnalysisData}(x, cls = NULL, reps = 1,
  pairwise = F, comparisons = list(), returnModels = F,
  seed = 1234, ...)
}
\description{
randomForest
}
