% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisResults.R
\name{raw}
\alias{raw}
\alias{raw<-}
\alias{raw,Analysis-method}
\alias{raw<-,Analysis-method}
\title{raw}
\usage{
raw(x)

raw(x) <- value

\S4method{raw}{Analysis}(x)

\S4method{raw}{Analysis}(x) <- value
}
\arguments{
\item{x}{S4 object of class Analysis}

\item{value}{S4 object of class AnalysisData}
}
\description{
Get or set an AnalysisData object from 
the raw slot of the Analysis class.
}
