% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/rawData-method.R
\docType{methods}
\name{rawData}
\alias{rawData}
\alias{rawData,Analysis-method}
\title{rawData-Analysis}
\usage{
rawData(x)

\S4method{rawData}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract raw data from an Analysis object
}
