% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/rawInfo-method.R
\docType{methods}
\name{rawInfo}
\alias{rawInfo}
\alias{rawData,Analysis-method}
\title{rawInfo-Analysis}
\usage{
rawInfo(x)

\S4method{rawData}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract raw info from an Analysis object
}
