% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{plan}
\alias{registerDoFuture}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{doFuture}{\code{\link[doFuture]{registerDoFuture}}}

  \item{future}{\code{\link[future]{plan}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

