% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeClasses}
\alias{removeClasses}
\alias{removeClasses,AnalysisData-method}
\alias{removeFeatures}
\alias{removeFeatures,AnalysisData-method}
\alias{removeSamples}
\alias{removeSamples,AnalysisData-method}
\title{Remove samples, classes or features}
\usage{
removeClasses(d, cls = "class", classes = c())

\S4method{removeClasses}{AnalysisData}(d, cls = "class", classes = c())

removeFeatures(d, features = character())

\S4method{removeFeatures}{AnalysisData}(d, features = character())

removeSamples(d, idx = "fileOrder", samples = c())

\S4method{removeSamples}{AnalysisData}(d, idx = "fileOrder", samples = c())
}
\arguments{
\item{d}{S4 object of class \code{AnalysisData}}

\item{cls}{info column to use for class information}

\item{classes}{classes to remove}

\item{features}{features to remove}

\item{idx}{info column containing sample indexes}

\item{samples}{sample indexes to remove}
}
\value{
An S4 object of class \code{AnalysisData} with samples, classes or features removed.
}
\description{
Exclusion of samples, classes or features from an \code{AnalysisData} object.
}
\section{Methods}{

\itemize{
\item \code{removeClasses}: Remove classes.
\item \code{removeFeatures}: Remove features.
\item \code{removeSamples}: Remove samples.
}
}

\examples{
library(metaboData)
 d <- analysisData(abr1$neg[,200:300],abr1$fact)
 
 ## Remove classes
 d \%>\% 
  removeClasses(cls = 'day',classes = 'H')
 
 ## Remove features
 d \%>\% 
  removeFeatures(features = c('N200','N201'))
 
 ## Remove samples
 d \%>\% 
  removeSamples(idx = 'injorder',samples = c(1,10))
}
