% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeMethods.R
\name{removeClasses}
\alias{removeClasses}
\alias{removeClasses,AnalysisData-method}
\title{removeClasses}
\usage{
removeClasses(d, cls = "class", classes = c())

\S4method{removeClasses}{AnalysisData}(d, cls = "class", classes = c())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class information}

\item{classes}{classes to remove}
}
\description{
Remove classes from an AnalysisData object.
}
