% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{removeCls}
\alias{removeCls}
\alias{removeCls,AnalysisData-method}
\alias{removeCls,Analysis-method}
\title{removeCls}
\usage{
removeCls(x, cls, ...)

\S4method{removeCls}{AnalysisData}(x, cls)

\S4method{removeCls}{Analysis}(x, cls, type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{cls}{name of sample information column to remove}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Remove a sample info column from a Analysis or AnalysisData object.
}
