% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeMethods.R
\name{removeFeatures}
\alias{removeFeatures}
\alias{removeFeatures,AnalysisData-method}
\title{removeFeatures}
\usage{
removeFeatures(d, features = character())

\S4method{removeFeatures}{AnalysisData}(d, features = character())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{features}{features to remove}
}
\description{
Remove features from an AnalysisData object.
}
