% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeMethods.R
\name{removeSamples}
\alias{removeSamples}
\alias{removeSamples,AnalysisData-method}
\title{removeSamples}
\usage{
removeSamples(d, idx = "fileOrder", samples = c())

\S4method{removeSamples}{AnalysisData}(d, idx = "fileOrder", samples = c())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{idx}{info column containing sample indexes}

\item{samples}{sample indexes to remove}
}
\description{
Remove samples from an AnalysisData object.
}
