% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeMethods.R
\docType{methods}
\name{removeVariables}
\alias{removeVariables}
\alias{removeVariables,AnalysisData-method}
\title{removeVariables}
\usage{
removeVariables(d, variables = character())

\S4method{removeVariables}{AnalysisData}(d, variables = character())
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{variables}{variables to remove}
}
\description{
Remove variables from an AnalysisData object.
}
