% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info.R
\name{replaceCls}
\alias{replaceCls}
\alias{replaceCls,AnalysisData-method}
\alias{replaceCls,Analysis-method}
\title{replaceCls}
\usage{
replaceCls(x, value, cls = "class", ...)

\S4method{replaceCls}{AnalysisData}(x, value, cls = "class")

\S4method{replaceCls}{Analysis}(x, value, cls = "class", type = "raw")
}
\arguments{
\item{x}{S4 object of class Analysis or AnalysisData}

\item{value}{vactor of new sample information for replacement}

\item{cls}{sample info column to replace}

\item{...}{arguments to pass to specific method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Replace a given sample info column from an Analysis or AnalysisData object.
}
