% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\name{response}
\alias{response}
\alias{response,RandomForest-method}
\title{response}
\usage{
response(x)

\S4method{response}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
Return the response variable name from a random forest analysis.
}
