% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{roc,RandomForest-method}
\alias{roc,list-method}
\alias{roc,Analysis-method}
\title{Receiver-operator characteristic (ROC) curves}
\usage{
roc(x)

\S4method{roc}{RandomForest}(x)

\S4method{roc}{list}(x)

\S4method{roc}{Analysis}(x)
}
\arguments{
\item{x}{S4 object of class \code{RandomForest}, \code{Analysis} or a list}

\item{idx}{sample information column to use for sample names. If \code{NULL}, the sample row number will be used. Sample names should be unique for each row of data.}
}
\value{
A tibble containing the ROC curves.
}
\description{
ROC curves for out-of-bag random forest predictions.
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day')

roc(rf)
}
