% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd.R
\name{rsd}
\alias{rsd}
\alias{rsd,AnalysisData-method}
\title{Calculate feature relative standard deviations}
\usage{
rsd(x, cls = "class")

\S4method{rsd}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{sample information column to use for class structure}
}
\value{
A tibble containing the computed RSD values.
}
\description{
Calculate relative standard deviation (RSD) percentage values for each
feature per class for a given sample information column.
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

rsd(d,cls = 'day')
}
