% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,Analysis-method}
\alias{show,Analysis-method}
\title{show-Analysis}
\usage{
\S4method{show}{Analysis}(object)
}
\arguments{
\item{object}{S4 object of class Analysis}
}
\description{
show method for Analysis class.
}
