% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,AnalysisData-method}
\alias{show,AnalysisData-method}
\title{show-AnalysisData}
\usage{
\S4method{show}{AnalysisData}(object)
}
\arguments{
\item{object}{S4 object of class AnalysisData}
}
\description{
show method for AnalysisData class
}
