% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,Data-method}
\alias{show,Data-method}
\title{show-Data}
\usage{
\S4method{show}{Data}(object)
}
\arguments{
\item{object}{S4 object of class Data}
}
\description{
show method for Data class
}
