% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,Univariate-method}
\alias{show,Univariate-method}
\title{show-Univariate}
\usage{
\S4method{show}{Univariate}(object)
}
\arguments{
\item{object}{S4 object of class Univariate}
}
\description{
Show method for the Univariate class.
}
