% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\name{sinfo}
\alias{sinfo}
\alias{sinfo,AnalysisData-method}
\alias{sinfo<-}
\title{sinfo}
\usage{
sinfo(x)

\S4method{sinfo}{AnalysisData}(x)

sinfo(x) <- value
}
\arguments{
\item{x}{S4 object of class Data}

\item{value}{tibble containing sample info}
}
\description{
Return sample info from an AnalysisData object.
}
