% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/split.R
\name{split}
\alias{split}
\alias{split,AnalysisData-method}
\title{split}
\usage{
split(x, cls = "class")

\S4method{split}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{sample information column to use for splitting}
}
\description{
Split an object of class AnalysisData into a list based 
given class information.
}
\examples{
library(metaboData)
d <- analysisData(abr1$neg,abr1$fact)
d <- split(d,cls = 'day')
}
