% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startAnalysis.R
\name{startAnalysis}
\alias{startAnalysis}
\title{startAnalysis}
\usage{
startAnalysis(data, info, params = parameters())
}
\description{
startAnalysis
}
\examples{
library(FIEmspro)
data(abr1)
p <- parameters()
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- startAnalysis(abr1$neg,abr1$fact,p) \%>\% 
preTreat 
}
