% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformArcSine}
\alias{transformArcSine}
\alias{transformArcSine,AnalysisData-method}
\title{transformArcSine}
\usage{
transformArcSine(dat)

\S4method{transformArcSine}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Arc-sine transformation of sample data.
}
