% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformAuto}
\alias{transformAuto}
\alias{transformAuto,AnalysisData-method}
\title{transformAuto}
\usage{
transformAuto(dat)

\S4method{transformAuto}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Auto scaling of sample data.
}
