% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformCenter}
\alias{transformCenter}
\alias{transformCenter,AnalysisData-method}
\title{transformCenter}
\usage{
transformCenter(dat)

\S4method{transformCenter}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Mean center sample data.
}
