% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformLevel}
\alias{transformLevel}
\alias{transformLevel,AnalysisData-method}
\title{transformLevel}
\usage{
transformLevel(d)

\S4method{transformLevel}{AnalysisData}(d)
}
\arguments{
\item{d}{S4 object of class AnalysisData}
}
\description{
Level scaling of sample data.
}
