% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\name{transformLn}
\alias{transformLn}
\alias{transformLn,AnalysisData-method}
\title{transformLn}
\usage{
transformLn(d, add = 1)

\S4method{transformLn}{AnalysisData}(d, add = 1)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{add}{value to add prior to transformation}
}
\description{
Natural logarithmic transformation of sample data.
}
