% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\name{transformLog10}
\alias{transformLog10}
\alias{transformLog10,AnalysisData-method}
\title{transformLog10}
\usage{
transformLog10(d, add = 1)

\S4method{transformLog10}{AnalysisData}(d, add = 1)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{add}{value to add prior to transformation}
}
\description{
Logarithmic transformation of sample data.
}
