% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformPareto}
\alias{transformPareto}
\alias{transformPareto,AnalysisData-method}
\title{transformPareto}
\usage{
transformPareto(dat)

\S4method{transformPareto}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Pareto scaling of sample data.
}
