% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformRange}
\alias{transformRange}
\alias{transformRange,AnalysisData-method}
\title{transformRange}
\usage{
transformRange(dat)

\S4method{transformRange}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Range scaling of sample data.
}
