% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformSQRT}
\alias{transformSQRT}
\alias{transformSQRT,AnalysisData-method}
\title{transformSQRT}
\usage{
transformSQRT(dat)

\S4method{transformSQRT}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Square root transformation of sample data.
}
