% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformVast}
\alias{transformVast}
\alias{transformVast,AnalysisData-method}
\title{transformVast}
\usage{
transformVast(dat)

\S4method{transformVast}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Vast scaling of sample data.
}
