% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{ttest}
\alias{ttest}
\alias{ttest,AnalysisData-method}
\title{Welch's t-test}
\usage{
ttest(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE
)

\S4method{ttest}{AnalysisData}(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE
)
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{vector of sample information column names to analyse}

\item{pAdjust}{p value adjustment method}

\item{comparisons}{named list of binary comparisons to analyse}

\item{returnModels}{should models be returned}
}
\value{
An S4 object of class \code{Univariate}.
}
\description{
Welch's t-test
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\% 
 keepClasses(cls = 'day',classes = c('H','5'))

## Perform t-test
ttest_analysis <- ttest(d,cls = 'day')

## Extract significant features
explanatoryFeatures(ttest_analysis)
}
