% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\alias{tune,AnalysisData-method}
\title{Tune random forest parameters}
\usage{
tune(
  x,
  cls = "class",
  mtry_range = floor(seq(mtry(x, cls = cls) - mtry(x, cls = cls)/2, mtry(x, cls = cls)
    + mtry(x, cls = cls)/2, length.out = 4)),
  ntree_range = 1000,
  seed = 1234
)

\S4method{tune}{AnalysisData}(
  x,
  cls = "class",
  mtry_range = floor(seq(mtry(x, cls = cls) - mtry(x, cls = cls)/2, mtry(x, cls = cls)
    + mtry(x, cls = cls)/2, length.out = 4)),
  ntree_range = 1000,
  seed = 1234
)
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{sample information column to use}

\item{mtry_range}{numeric vector of \code{mtry} values to search}

\item{ntree_range}{numeric vector of \code{ntree} values to search}

\item{seed}{random number seed}
}
\value{
A list containing the optimal \code{mtry} and \code{ntree} parameters.
This is suitable for use as the \code{rf} argument in method \code{randomForest()}.
}
\description{
Tune the \code{mtry} and \code{ntree} random forest parameters using a grid search approach.
}
\details{
Parameter tuning is performed by grid search of all combinations of the \code{mtry_range} and \code{ntree_range} vectors provided.
The optimal parameter values are selected using the out-of-bag error estimates of the \code{margin} metric for classification and the \code{rmse} (root-mean-square error) metric for regression.
}
\examples{
library(metaboData)

## Prepare some data
x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
  occupancyMaximum(cls = 'day') \%>\%
  transformTICnorm()

## Tune the `mtry` parameter for the `day` response
tune(x,cls = 'day')
}
