% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\name{type}
\alias{type}
\alias{type,RandomForest-method}
\title{type}
\usage{
type(x)

\S4method{type}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
Return the random forest analysis type.
}
