% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Univariate-class}
\alias{Univariate-class}
\title{Univariate}
\description{
An S4 class for univariate test models and results.
}
\section{Slots}{

\describe{
\item{\code{type}}{univariate test type}

\item{\code{response}}{response}

\item{\code{models}}{list of model objects}

\item{\code{results}}{tibble containing test results}
}}

