% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/aggregateMethods.R
\docType{methods}
\name{aggregateMedian}
\alias{aggregateMedian}
\alias{aggregateMedian,Data-method}
\title{aggregateMedian}
\usage{
aggregateMedian(dat, cls = "class")

\S4method{aggregateMedian}{Data}(dat, cls = "class")
}
\arguments{
\item{dat}{S4 object of class Data}

\item{cls}{info column to use for class data}
}
\description{
Median aggregation of sample data.
}
