% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/aggregateMethods.R
\docType{methods}
\name{aggregateSum}
\alias{aggregateSum}
\alias{aggregateSum,AnalysisData-method}
\title{aggregateSum}
\usage{
aggregateSum(dat, cls = "class")

\S4method{aggregateSum}{AnalysisData}(dat, cls = "class")
}
\arguments{
\item{dat}{S4 object of class Data}

\item{cls}{info column to use for class data}
}
\description{
Sum aggregation of sample data.
}
