% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Analysis-class}
\alias{Analysis-class}
\title{Analysis}
\description{
An S4 class to store analysis results
}
\section{Slots}{

\describe{
\item{\code{log}}{list containing analysis dates and time}

\item{\code{parameters}}{class AnalysisParameters containing the analysis parameters}

\item{\code{raw}}{list containing info and raw data}

\item{\code{pre-treated}}{list containing preTreated info and raw data}

\item{\code{modelling}}{list containing modelling results}

\item{\code{correlations}}{tibble containing weighted edgelist of correlations}
}}

