% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AnalysisData-class}
\alias{AnalysisData-class}
\title{AnalysisData}
\description{
An S4 class for sample data and info.
}
\section{Slots}{

\describe{
\item{\code{data}}{sample data}

\item{\code{info}}{sample info}
}}

