% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AnalysisParameters-class}
\alias{AnalysisParameters-class}
\title{AnalysisParameters S4 class}
\description{
An S4 class to store analysis parameters.
}
\section{Slots}{

\describe{
\item{\code{pre-treatment}}{list containing parameters for data pre-treatment}

\item{\code{modelling}}{list containing parameters for modelling}

\item{\code{correlations}}{list containing parameters for correlations}
}}

