% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisParameters.R
\name{analysisParameters}
\alias{analysisParameters}
\title{analysisParameters}
\usage{
analysisParameters(elements = c("preTreat", "modelling", "correlations"))
}
\arguments{
\item{elements}{character vector containing elements for analysis (see Details). Default includes all available elements.}
}
\description{
Initiate default analysis parameters for analysis elements.
}
\details{
Analysis elements can include:
\itemize{
\item preTreat
\item classification
\item featureSelection
\item correlations
}
}
