% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{analysisParameters}
\alias{analysisParameters}
\title{analysisParameters}
\usage{
analysisParameters(elements = analysisElements())
}
\arguments{
\item{elements}{character vector containing elements for analysis. 
Default includes all available elements from \code{analysisElements}.}
}
\description{
Initiate default analysis parameters for analysis elements.
}
