% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/changeParameter.R
\name{changeParameter}
\alias{changeParameter}
\alias{changeParameter,AnalysisParameters-method}
\title{changeParameter}
\usage{
changeParameter(
  x,
  parameterName,
  newValue,
  elements = c("preTreat", "modelling", "correlations")
)

\S4method{changeParameter}{AnalysisParameters}(x, parameterName, newValue, elements = analysisElements())
}
\arguments{
\item{parameterName}{Name of the parameter to change}

\item{newValue}{New value of the parameter}

\item{elements}{Character vector of analysis elements to target parameter change. Can be any returned by \code{analysisElements}.}

\item{parameters}{S4 object of class AnalysisParameters in which to change parameters}
}
\description{
change analysis parameters
}
\details{
For the parameter name selected, all parameters with that name will be altered.
To individually change identically named parameters use the \code{@} operator to access the appropriate slot directly.
}
\examples{
p <- analysisParameters()
p <- changeParameter(p,'clusterType','PSOCK')
}
