% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeParameter.R
\name{changeParameter}
\alias{changeParameter}
\title{changeParameter}
\usage{
changeParameter(parameterName, newValue, parameters)
}
\arguments{
\item{parameterName}{Name of the parameter to change}

\item{newValue}{New value of the parameter}

\item{parameters}{S4 object of class AnalysisParameters in which to change parameters}
}
\description{
change analysis parameters
}
\details{
For the parameter name selected, all parameters with that name will be altered.
To individually change identically named parameters use the \code{@} operator to access the appropriate slot directly.
}
\examples{
p <- analysisParameters()
p <- changeParameter('clusterType','PSOCK',p)
}
