% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/correctionMethods.R
\name{correctionCenter}
\alias{correctionCenter}
\alias{correctionCenter,AnalysisData-method}
\title{correctionCenter}
\usage{
correctionCenter(
  d,
  block = "block",
  type = "median",
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)

\S4method{correctionCenter}{AnalysisData}(
  d,
  block = "block",
  type = "median",
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{block}{info column containing sample block groupings to use for correction}

\item{type}{averaging to use; eg. mean or median}

\item{nCores}{number of cores for parallisation}

\item{clusterType}{cluster type for parallisation}
}
\description{
Batch correction using average centering.
}
