% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\alias{correlations,AnalysisData-method}
\alias{correlations,Analysis-method}
\title{correlations}
\usage{
correlations(d, ...)

\S4method{correlations}{AnalysisData}(
  d,
  method = "pearson",
  pAdjustMethod = "bonferroni",
  corPvalue = 0.05
)

\S4method{correlations}{Analysis}(d)
}
\arguments{
\item{d}{S4 object of class \code{AnalysisData} or \code{Analysis}}

\item{...}{arguments to pass to specific method}

\item{method}{correlation method. One of \code{pearson} or \code{spearman}.}

\item{pAdjustMethod}{p-value adjustment method. See \code{?p.adjust} for available methods.}

\item{corPvalue}{p-value cut-off threshold for significance}
}
\description{
Feature correlation analysis.
}
\details{
Correlation analyses can be used to identify associated features within data sets.
This can be useful to identifying clusters of related features that can be used to annotate metabolites within data sets.
All features are compared and the returned table of correlations are p-value thresholded using the specified cut-off.
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

correlations(d)
}
