% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\name{dat}
\alias{dat}
\alias{dat,AnalysisData-method}
\alias{dat<-}
\title{dat}
\usage{
dat(x)

\S4method{dat}{AnalysisData}(x)

dat(x) <- value
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{value}{tibble containing sample data}
}
\description{
Return or sest sample data in an AnalysisData object.
}
