% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisResults.R
\name{dat}
\alias{dat}
\alias{dat<-}
\alias{dat,AnalysisData-method}
\alias{dat<-,AnalysisData-method}
\alias{dat,Analysis-method}
\alias{dat<-,Analysis-method}
\title{dat}
\usage{
dat(x, ...)

dat(x, ...) <- value

\S4method{dat}{AnalysisData}(x)

\S4method{dat}{AnalysisData}(x) <- value

\S4method{dat}{Analysis}(x, type = "pre-treated")

\S4method{dat}{Analysis}(x, type = "pre-treated") <- value
}
\arguments{
\item{x}{S4 object of class AnalysisData or Analysis}

\item{...}{arguments to pass to the appropriate method}

\item{value}{tibble containing sample data}

\item{type}{data type to extract or set. 
Should be one of "raw" or "pre-treated"}
}
\description{
Return or set sample data in an 
AnalysisData or Analysis objects.
}
