% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modelling.R
\docType{methods}
\name{explanatoryFeatures}
\alias{explanatoryFeatures}
\alias{explanatoryFeatures,Univariate-method}
\alias{explanatoryFeatures,RandomForest-method}
\title{explanatoryFeatures}
\usage{
explanatoryFeatures(x, ...)

\S4method{explanatoryFeatures}{Univariate}(x, threshold = 0.05)

\S4method{explanatoryFeatures}{RandomForest}(x,
  measure = "FalsePositiveRate", threshold = 0.05)
}
\arguments{
\item{x}{S4 object of class RandomForest or Univariate}

\item{...}{arguments to parse to method for specific class}
}
\description{
Extract explanatory features from modelling results.
}
