% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/imputeMethods.R
\docType{methods}
\name{imputeAll}
\alias{imputeAll}
\alias{imputeAll,AnalysisData-method}
\title{imputeAll}
\usage{
imputeAll(d, occupancy = 2/3, parallel = "variables",
  nCores = detectCores() * 0.75, clusterType = getClusterType(),
  seed = 1234)

\S4method{imputeAll}{AnalysisData}(d, occupancy = 2/3,
  parallel = "variables", nCores = detectCores() * 0.75,
  clusterType = getClusterType(), seed = 1234)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{occupancy}{occupancy threshold for imputation}

\item{parallel}{parallel type to use. See `?missForest` for details}

\item{nCores}{number of cores for parallisation}

\item{clusterType}{cluster type for parallisation}

\item{seed}{random number seed}
}
\description{
Impute missing values across all samples using Random Forest.
}
