% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisData.R
\docType{methods}
\name{info}
\alias{info}
\alias{info,Data-method}
\title{info}
\usage{
info(x)

\S4method{info}{Data}(x)
}
\arguments{
\item{x}{S4 object of class Data}
}
\description{
Return sample info from a Data object.
}
