% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/univariateMethods.R
\docType{methods}
\name{linearRegression}
\alias{linearRegression}
\alias{linearRegression,AnalysisData-method}
\title{linearRegression}
\usage{
linearRegression(x, cls = "class", pAdjust = "bonferroni",
  returnModels = F)

\S4method{linearRegression}{AnalysisData}(x, cls = "class",
  pAdjust = "bonferroni", returnModels = F)
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{vector of sample info column names to regress}

\item{pAdjust}{p value adjustment method}

\item{returnModels}{should models be returned}
}
\description{
Linear regression
}
