% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/randomForest.R
\name{measures}
\alias{measures}
\alias{measures,RandomForest-method}
\alias{measures,list-method}
\title{measures}
\usage{
measures(x)

\S4method{measures}{RandomForest}(x)

\S4method{measures}{list}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}
}
\description{
return measures results from a RandomForest object
}
