% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolyse.R
\name{metabolyse}
\alias{metabolyse}
\title{metabolyse}
\usage{
metabolyse(data, info, params = analysisParameters())
}
\description{
metabolyse
}
\examples{
library(FIEmspro)
data(abr1)
p <- analysisParameters('preTreat')
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)  
}
