% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolyse.R
\name{metabolyse}
\alias{metabolyse}
\title{metabolyse}
\usage{
metabolyse(data, info, parameters = analysisParameters(), verbose = TRUE)
}
\arguments{
\item{data}{tibble or data.frame containing data to analyse}

\item{info}{tibble or data.frame containing data info or meta data}

\item{parameters}{an object of AnalysisParameters class containing 
parameters for analysis. Default calls \code{analysisParameters()}}

\item{verbose}{should output be printed to the console}
}
\description{
Analyse data based on specified analysis elements.
}
\examples{
library(metaboData)

## Generate analysis parameters
p <- analysisParameters()

## Alter pre-treatment and modelling parameters to use different methods
parameters(p,'pre-treatment') <- preTreatmentParameters(
  list(occupancyFilter = 'maximum',
       transform = 'TICnorm')
)
parameters(p,'modelling') <- modellingParameters('anova')

## Change "cls" and "nCores" parameters 
changeParameter(p,'cls') <- 'day'
changeParameter(p,'nCores') <- 2

## Run analysis using a subset of the abr1 negative mode data set
analysis <- metabolyse(abr1$neg[,1:200],
                       abr1$fact,
                       p)
}
\seealso{
\linkS4class{AnalysisParameters} \link{analysisParameters}
}
