% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolyse.R
\name{metabolyse}
\alias{metabolyse}
\title{metabolyse}
\usage{
metabolyse(data, info, parameters = analysisParameters(), verbose = T)
}
\arguments{
\item{data}{tibble or data.frame containing data to analyse}

\item{info}{tibble or data.frame containing data info or meta data}

\item{parameters}{an object of AnalysisParameters class containing parameters for analysis. Default calls \code{analysisParameters()}}
}
\description{
Analyse data based on specified analysis elements.
}
\examples{
library(metaboData)
data(abr1)
p <- analysisParameters('preTreat')
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)  
}
\seealso{
\linkS4class{AnalysisParameters} \link{analysisParameters}
}
