% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{modellingParameters}
\alias{modellingParameters}
\title{modellingParameters}
\usage{
modellingParameters(methods)
}
\arguments{
\item{methods}{character vector of available methods. 
Use \code{modellingMethods()} to see available methods.}
}
\description{
Return default parameters for a given modelling method.
}
\examples{
p <- analysisParameters()
parameters(p,'modelling') <- modellingParameters('anova')
}
