% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/occupancyMethods.R
\name{occupancy}
\alias{occupancy}
\alias{occupancy,AnalysisData-method}
\title{occupancy}
\usage{
occupancy(x, cls = "class")

\S4method{occupancy}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{info column to use for class data}
}
\description{
Return tibble containg proportional class occupancy for each feature for a given class info column.
}
