% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\name{occupancy}
\alias{occupancy}
\alias{occupancy,AnalysisData-method}
\title{occupancy}
\usage{
occupancy(d, cls = "class")

\S4method{occupancy}{AnalysisData}(d, cls = "class")
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class data}
}
\description{
Return tibble containg proportional class occupancy
for each feature for a given class info column.
}
