% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Parameters-class}
\alias{Parameters-class}
\title{Parameters}
\description{
An S4 class to store analysis parameters
}
\section{Slots}{

\describe{
\item{\code{preTreat}}{list containing parameters for data pre-treatment}

\item{\code{classification}}{list containing parameters for classification}

\item{\code{featureSelection}}{list containing parameters for feature selection}

\item{\code{correlations}}{list containing parameters for correlations}
}}

