% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters<-}
\alias{parameters,Analysis-method}
\alias{parameters<-,Analysis-method}
\alias{parameters,AnalysisParameters-method}
\alias{parameters<-,AnalysisParameters-method}
\title{parameters}
\usage{
parameters(x, ...)

parameters(x, element) <- value

\S4method{parameters}{Analysis}(x)

\S4method{parameters}{Analysis}(x) <- value

\S4method{parameters}{AnalysisParameters}(x, element)

\S4method{parameters}{AnalysisParameters}(x, element) <- value
}
\arguments{
\item{x}{S4 object of class Analysis}

\item{...}{arguments to pass to the appropriate method}

\item{element}{analysis element for parameters to extract or assign. 
Should be one of those returned by \code{analysisElements()}}

\item{value}{list containing parameter values}
}
\description{
Extract parameters from an Analysis object or 
extract or set parameters in an AnalysisParameters object.
}
\examples{
p <- analysisParameters()

## extract pre-treatment parameters
parameters(p,'pre-treatment')

## set pre-treatment parameters
parameters(p,'pre-treatment') <- preTreatmentParameters(
  list(
    remove = 'classes',
    QC = c('RSDfilter','removeQC'),
    transform = 'TICnorm'
  )
)
}
