% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/plotExplanatoryHeatmap-method.R
\name{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap,Univariate-method}
\alias{plotExplanatoryHeatmap,RandomForest-method}
\alias{plotExplanatoryHeatmap,list-method}
\alias{plotExplanatoryHeatmap,Analysis-method}
\title{plotExplanatoryHeatmap}
\usage{
plotExplanatoryHeatmap(x, ...)

\S4method{plotExplanatoryHeatmap}{Univariate}(
  x,
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  dendrogram = TRUE
)

\S4method{plotExplanatoryHeatmap}{RandomForest}(
  x,
  metric = "FalsePositiveRate",
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  dendrogram = TRUE
)

\S4method{plotExplanatoryHeatmap}{list}(
  x,
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE
)

\S4method{plotExplanatoryHeatmap}{Analysis}(
  x,
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE
)
}
\arguments{
\item{x}{object of class Univariate, RandomForest or 
Analysis containing modelling results}

\item{...}{arguments to pass to the appropriate method}

\item{threshold}{score threshold to use for specifying explantory features}

\item{distanceMeasure}{distance measure to use for clustering. See details.}

\item{clusterMethod}{clustering method to use. See details}

\item{featureNames}{should feature names be plotted?}

\item{dendrogram}{TRUE/FALSE. Should the dendrogram be plotted?}

\item{metric}{importance metric on which to retrieve explanatory feautres}
}
\description{
plot a heatmap of explanatory features
}
\details{
Options for distance measures are as for \code{dist()}.
Clustering methods are as given for \code{hclust()}.
}
\examples{
\dontrun{
library(metaboData)
x <- analysisData(data = abr1$neg,info = abr1$fact)

## random forest classification example
random_forest <- randomForest(x,cls = 'day')

plotExplanatoryHeatmap(random_forest)

## random forest regression example
random_forest <- randomForest(x,cls = 'injorder')

plotExplanatoryHeatmap(random_forest,metric = '\%IncMSE',threshold = 2)
}
}
\seealso{
\link{dist} \link{hclust}
}
