% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExplanatoryHeatmap.R
\name{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap}
\alias{plotExplanatoryHeatmap,Univariate-method}
\alias{plotExplanatoryHeatmap,RandomForest-method}
\alias{plotExplanatoryHeatmap,list-method}
\alias{plotExplanatoryHeatmap,Analysis-method}
\title{Heatmap plot of explantory features}
\usage{
plotExplanatoryHeatmap(x, ...)

\S4method{plotExplanatoryHeatmap}{Univariate}(
  x,
  threshold = 0.05,
  title = "",
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  dendrogram = TRUE,
  featureLimit = Inf
)

\S4method{plotExplanatoryHeatmap}{RandomForest}(
  x,
  metric = "FalsePositiveRate",
  threshold = 0.05,
  title = "",
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  dendrogram = TRUE,
  featureLimit = Inf
)

\S4method{plotExplanatoryHeatmap}{list}(
  x,
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  featureLimit = Inf
)

\S4method{plotExplanatoryHeatmap}{Analysis}(
  x,
  threshold = 0.05,
  distanceMeasure = "euclidean",
  clusterMethod = "ward.D2",
  featureNames = TRUE,
  featureLimit = Inf
)
}
\arguments{
\item{x}{object of class \code{Univariate}, \code{RandomForest} or
\code{Analysis}}

\item{...}{arguments to pass to the appropriate method}

\item{threshold}{score threshold to use for specifying explanatory features}

\item{title}{plot title}

\item{distanceMeasure}{distance measure to use for clustering. See details.}

\item{clusterMethod}{clustering method to use. See details}

\item{featureNames}{should feature names be plotted?}

\item{dendrogram}{TRUE/FALSE. Should the dendrogram be plotted?}

\item{featureLimit}{The maximum number of features to plot}

\item{metric}{importance metric on which to retrieve explanatory features}
}
\description{
Plot a heatmap of explanatory features.
}
\details{
Distance measures can be one of any that can be used for the \code{method} argument of \code{\link[=dist]{dist()}}.

Cluster methods can be one of any that can be used for the \code{method} argument of \code{\link[=hclust]{hclust()}}.
}
\examples{
library(metaboData)
x <- analysisData(data = abr1$neg[,200:300],info = abr1$fact)

## random forest classification example
random_forest <- randomForest(x,cls = 'day')

plotExplanatoryHeatmap(random_forest)

## random forest regression example
random_forest <- randomForest(x,cls = 'injorder')

plotExplanatoryHeatmap(random_forest,metric = '\%IncMSE',threshold = 2)
}
