% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotFeature-method.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,Analysis-method}
\title{plotFeature}
\usage{
plotFeature(analysis, feature, cls = "class")

\S4method{plotFeature}{Analysis}(analysis, feature, cls = "class")
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{feature}{feature to plot}

\item{cls}{info column to use for class labels}
}
\description{
Plot a feature trend.
}
\examples{
library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotFeature(analysis,'N133',cls = 'day')
}
