% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotFeature-method.R
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,Analysis-method}
\alias{plotFeature,AnalysisData-method}
\title{plotFeature}
\usage{
plotFeature(analysis, feature, cls = "class", label = NULL, labelSize = 2, ...)

\S4method{plotFeature}{Analysis}(
  analysis,
  feature,
  cls = "class",
  label = NULL,
  labelSize = 2,
  type = "pre-treated"
)

\S4method{plotFeature}{AnalysisData}(analysis, feature, cls = "class", label = NULL, labelSize = 2)
}
\arguments{
\item{analysis}{object of class Analysis, AnalysisData, Univariate 
or RandomForest containing analysis results}

\item{feature}{feature to plot}

\item{cls}{info column to use for class labels}

\item{label}{info column to use for sample labels}

\item{labelSize}{sample label size}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{pre-treated} data to plot}
}
\description{
Plot a feature trend.
}
\examples{
\dontrun{
library(metaboData)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotFeature(analysis,'N133',cls = 'day')
}
}
