% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotImportance}
\alias{plotImportance}
\alias{plotImportance,Univariate-method}
\alias{plotImportance,RandomForest-method}
\alias{plotImportance,list-method}
\title{plotImportance}
\usage{
plotImportance(x, ...)

\S4method{plotImportance}{Univariate}(x, response = "class", rank = TRUE, threshold = 0.05)

\S4method{plotImportance}{RandomForest}(x, metric = "FalsePositiveRate", rank = TRUE)

\S4method{plotImportance}{list}(x, metric = "FalsePositiveRate")
}
\arguments{
\item{x}{S4 object of class Univariate or RandomForest}

\item{...}{arguments to pass to specific method}

\item{response}{Response results to plot}

\item{rank}{Rank feature order for plotting}

\item{threshold}{explanatory threshold line for the output plot}

\item{metric}{Importance metric to plot}
}
\description{
Plot univariate or random forest feature importance.
}
