% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDA.R
\name{plotLDA}
\alias{plotLDA}
\alias{plotLDA,AnalysisData-method}
\alias{plotLDA,Analysis-method}
\title{Principle Component - Linear Discriminant Analysis plot}
\usage{
plotLDA(
  analysis,
  cls = "class",
  label = NULL,
  scale = TRUE,
  center = TRUE,
  xAxis = "DF1",
  yAxis = "DF2",
  shape = FALSE,
  ellipses = TRUE,
  title = "PC-LDA",
  legendPosition = "bottom",
  labelSize = 2,
  ...
)

\S4method{plotLDA}{AnalysisData}(
  analysis,
  cls = "class",
  label = NULL,
  scale = TRUE,
  center = TRUE,
  xAxis = "DF1",
  yAxis = "DF2",
  shape = FALSE,
  ellipses = TRUE,
  title = "PC-LDA",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotLDA}{Analysis}(
  analysis,
  cls = "class",
  label = NULL,
  scale = TRUE,
  center = TRUE,
  xAxis = "DF1",
  yAxis = "DF2",
  shape = FALSE,
  ellipses = TRUE,
  title = "PC-LDA",
  legendPosition = "bottom",
  labelSize = 2,
  type = "raw"
)
}
\arguments{
\item{analysis}{S4 object of class \code{AnalysisData} or \code{Analysis}}

\item{cls}{name of sample information column to use for class labels}

\item{label}{name of sample information column to use for sample labels. Set to NULL for no labels.}

\item{scale}{scale the data}

\item{center}{center the data}

\item{xAxis}{principle component to plot on the x-axis}

\item{yAxis}{principle component to plot on the y-axis}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points.
Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\\%
confidence ellipses for each class}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument
of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{pre-treated} data to plot}
}
\description{
Plot linear discriminant analysis results of pre-treated data
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg,abr1$fact) \%>\% 
 occupancyMaximum(cls = 'day')

## LDA plot
plotLDA(d,cls = 'day')
}
