% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotLDA-method.R
\name{plotLDA}
\alias{plotLDA}
\alias{plotLDA,AnalysisData-method}
\alias{plotLDA,Analysis-method}
\title{plotLDA}
\usage{
plotLDA(
  analysis,
  cls = "class",
  label = NULL,
  scale = T,
  center = T,
  xAxis = "DF1",
  yAxis = "DF2",
  ellipses = T,
  title = "Principle Component - Linear Discriminant Analysis (PC-LDA) plot",
  legend = TRUE,
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotLDA}{AnalysisData}(
  analysis,
  cls = "class",
  label = NULL,
  scale = T,
  center = T,
  xAxis = "DF1",
  yAxis = "DF2",
  ellipses = T,
  title = "Principle Component - Linear Discriminant Analysis (PC-LDA)",
  legend = TRUE,
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotLDA}{Analysis}(
  analysis,
  cls = "class",
  label = NULL,
  scale = T,
  center = T,
  xAxis = "DF1",
  yAxis = "DF2",
  ellipses = T,
  title = "Principle Component - Linear Discriminant Analysis (PC-LDA)",
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{analysis}{object of class Analysis or AnalysisData containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{scale}{scale the data}

\item{center}{center the data}

\item{xAxis}{principle component to plot on the x-axis}

\item{yAxis}{principle component to plot on the y-axis}

\item{ellipses}{should multivariate normal distribution 95\% confidence ellipses be plotted for each class?}

\item{title}{plot title}

\item{legend}{TRUE/FALSE should a legend be plotted. Useful for many classes. Defaults to TRUE.}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}. Ignored if \code{legend = FALSE}.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot linear discriminant analysis resultus of pre-treated data
}
\examples{

library(metaboData)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotLDA(analysis,cls = 'day')
}
