% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\docType{methods}
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,RandomForest-method}
\title{plotMDS}
\usage{
plotMDS(x, cls = "class", label = NULL, ellipses = T, title = "",
  legendPosition = "bottom", labelSize = 2)

\S4method{plotMDS}{RandomForest}(x, cls = "class", label = NULL,
  ellipses = T, title = "", legendPosition = "bottom",
  labelSize = 2)
}
\description{
plotMDS
}
