% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,RandomForest-method}
\title{plotMDS}
\usage{
plotMDS(
  x,
  cls = "class",
  label = NULL,
  ellipses = T,
  title = "",
  legend = TRUE,
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotMDS}{RandomForest}(
  x,
  cls = "class",
  label = NULL,
  ellipses = T,
  title = "",
  legend = TRUE,
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{cls}{info column to use for sample labelling, Set to NULL for no labelling.}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{ellipses}{should multivariate normal distribution 95\% confidence ellipses be plotted for each class?}

\item{title}{plot title}

\item{legend}{TRUE/FALSE should a legend be plotted. Useful for many classes. Defaults to TRUE.}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}. Ignored if \code{legend = FALSE}.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot multidimensional scaling plot for a RandomForest object.
}
