% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,RandomForest-method}
\alias{plotMDS,list-method}
\title{plotMDS}
\usage{
plotMDS(
  x,
  cls = "class",
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotMDS}{RandomForest}(
  x,
  cls = "class",
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotMDS}{list}(
  x,
  cls = "class",
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{cls}{info column to use for sample labelling, 
Set to NULL for no labelling.}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points. 
Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\%
confidence ellipses for each class}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument 
of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
Plot multidimensional scaling plot for a RandomForest object.
}
