% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\docType{methods}
\name{plotMeasures}
\alias{plotMeasures}
\alias{plotMeasures,RandomForest-method}
\title{plotMeasures}
\usage{
plotMeasures(x, predictor = "class")

\S4method{plotMeasures}{RandomForest}(x, predictor = "class")
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{predictor}{predictor results to plot}
}
\description{
Plot random forest model measures.
}
