% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotMeasures}
\alias{plotMeasures}
\alias{plotMeasures,RandomForest-method}
\title{plotMeasures}
\usage{
plotMeasures(x, response = "class")

\S4method{plotMeasures}{RandomForest}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{response}{response results to plot}
}
\description{
Plot random forest model measures.
}
