% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotMetrics}
\alias{plotMetrics}
\alias{plotMetrics,RandomForest-method}
\alias{plotMetrics,list-method}
\title{plotMetrics}
\usage{
plotMetrics(x, response = "class")

\S4method{plotMetrics}{RandomForest}(x)

\S4method{plotMetrics}{list}(x)
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{response}{response results to plot}
}
\description{
Plot random forest model metrics.
}
