% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOccupancy.R
\name{plotOccupancy}
\alias{plotOccupancy}
\alias{plotOccupancy,AnalysisData-method}
\alias{plotOccupancy,Analysis-method}
\title{Plot class occupancy distributions}
\usage{
plotOccupancy(x, cls = "class", ...)

\S4method{plotOccupancy}{AnalysisData}(x, cls = "class")

\S4method{plotOccupancy}{Analysis}(x, cls = "class", type = "raw")
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData} or \code{Analysis}}

\item{cls}{sample information column to use for class labels}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{preTreated} data to plot}
}
\description{
Plot class occupancy distributions.
}
\examples{
#' library(metaboData)

d <- analysisData(abr1$neg,abr1$fact)

## Plot class occupancy distributions
plotOccupancy(d,cls = 'day')
}
