% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotPCA-method.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,Analysis-method}
\title{plotPCA}
\usage{
plotPCA(analysis, cls = "class", scale = T, center = T,
  xAxis = "PC1", yAxis = "PC2")

\S4method{plotPCA}{Analysis}(analysis, cls = "class", scale = T,
  center = T, xAxis = "PC1", yAxis = "PC2")
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{scale}{scale the data}

\item{center}{center the data}

\item{xAxis}{principle component to plot on the x-axis}

\item{yAxis}{principle component to plot on the y-axis}
}
\description{
Plot principle component analysis results of pre-treated data.
}
\examples{

library(FIEmspro)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotPCA(analysis,cls = 'day')
}
