% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotROC}
\alias{plotROC}
\alias{plotROC,RandomForest-method}
\title{plotROC}
\usage{
plotROC(x, title = "", legendPosition = "bottom")

\S4method{plotROC}{RandomForest}(x, title = "", legendPosition = "bottom")
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}. Set to "none" to remove legend.}
}
\description{
plot reciever operator characteristic curves for a RandomForest object.
}
