% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/modellingPlots.R
\name{plotROC}
\alias{plotROC}
\alias{plotROC,RandomForest-method}
\title{plotROC}
\usage{
plotROC(x, title = "")

\S4method{plotROC}{RandomForest}(x, title = "")
}
\arguments{
\item{x}{S4 object of class RandomForest}

\item{title}{plot title}
}
\description{
plot reciever operator characteristic curves for a RandomForest object.
}
