% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotRSD-method.R
\docType{methods}
\name{plotRSD}
\alias{plotRSD}
\alias{plotRSD,Analysis-method}
\title{plotRSD}
\usage{
plotRSD(analysis, cls = "class", QCidx = "QC", QCparameters = NULL,
  modes = T, histBins = 30)

\S4method{plotRSD}{Analysis}(analysis, cls = "class", QCidx = "QC",
  QCparameters = NULL, modes = T, histBins = 30)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{QCparameters}{alternative parameters for QC sample pre-treatment. See details}

\item{modes}{split modes if present}

\item{histBins}{number of bins to use for histogram plotting}
}
\description{
Plot RSD distributions of raw data in quality control samples.
}
\details{
If QCparameters is set as \code{NULL}, the default QC pre-treatment parameters are used as given by \code{analysisParameters('preTreat')}. Alternative pre-treatment routines can be used by specifying an \code{AnalysisParameters} object for \code{QCparameters}.
}
\examples{
\dontrun{

files <- list.files(
  system.file(
    'DataSets/FIE-HRMS/BdistachyonEcotypes',
    package = 'metaboData'),
  full.names = TRUE)

info <- readr::read_csv(files[grepl('runinfo',files)])
files <- files[!grepl('runinfo',files)]

binDat <- binneR::binneRlyse(files, 
                       info, 
                       parameters = binneR::binParameters())

p <- new('AnalysisParameters')

analysis <- metabolyse(dplyr::bind_cols(binnedData(binDat)),info(binDat),p)

plotRSD(analysis)
}
}
