% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRSD.R
\name{plotRSD}
\alias{plotRSD}
\alias{plotRSD,AnalysisData-method}
\alias{plotRSD,Analysis-method}
\title{Plot RSD distributions}
\usage{
plotRSD(analysis, cls = "class", ...)

\S4method{plotRSD}{AnalysisData}(analysis, cls = "class")

\S4method{plotRSD}{Analysis}(
  analysis,
  cls = "class",
  QCidx = "QC",
  QCparameters = NULL,
  histBins = 30,
  title = ""
)
}
\arguments{
\item{analysis}{object of class \code{AnalysisData} or \code{Analysis}}

\item{cls}{information column to use for class labels}

\item{...}{arguments to pass to the appropriate method}

\item{QCidx}{QC sample label}

\item{QCparameters}{alternative parameters for QC sample pre-treatment.
See details}

\item{histBins}{number of bins to use for histogram plotting}

\item{title}{plot title}
}
\description{
Plot RSD distributions of raw data in quality control samples.
}
\examples{
#' library(metaboData)

d <- analysisData(abr1$neg,abr1$fact)

## Plot class RSD distributions
plotRSD(d,cls = 'day')
}
