% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSupervisedRF-method.R
\docType{methods}
\name{plotSupervisedRF}
\alias{plotSupervisedRF}
\alias{plotSupervisedRF,Analysis-method}
\title{plotSupervisedRF}
\usage{
plotSupervisedRF(analysis, cls = "class", label = "sampleName", ...)

\S4method{plotSupervisedRF}{Analysis}(analysis, cls = "class",
  label = "name", seed = 1234, ...)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for sample classes}

\item{label}{info column to use for sample labels, Set to NULL for no labels.}

\item{...}{additional parameters to pass to randomForest}

\item{seed}{random number seed}
}
\description{
plotSupervisedRF
}
