% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSupervisedRF-method.R
\name{plotSupervisedRF}
\alias{plotSupervisedRF}
\alias{plotSupervisedRF,AnalysisData-method}
\alias{plotSupervisedRF,Analysis-method}
\title{plotSupervisedRF}
\usage{
plotSupervisedRF(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  ROC = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotSupervisedRF}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  ROC = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotSupervisedRF}{Analysis}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = T,
  ROC = T,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{x}{object of class Analysis or AnalysisData containing analysis results}

\item{cls}{info column to use for sample classes}

\item{rf}{list of additional parameters to pass to randomForest}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points. Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\% confidence ellipses for each class}

\item{ROC}{should reciever-operator characteristics be plotted?}

\item{seed}{random number seed}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
plotSupervisedRF
}
\examples{
\dontrun{
library(metaboData)
data(abr1)
p <- analysisParameters('preTreat')
p@preTreat <- list(
  occupancyFilter = list(maximum = list()),
  transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)  

plotSupervisedRF(analysis,cls = 'day',label = 'name')
}
}
