% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSupervisedRF.R
\name{plotSupervisedRF}
\alias{plotSupervisedRF}
\alias{plotSupervisedRF,AnalysisData-method}
\alias{plotSupervisedRF,Analysis-method}
\title{Supervised random forest MDS plot}
\usage{
plotSupervisedRF(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  ROC = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2,
  ...
)

\S4method{plotSupervisedRF}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  ROC = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotSupervisedRF}{Analysis}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  ROC = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2,
  type = c("pre-treated", "raw")
)
}
\arguments{
\item{x}{object of class \code{AnalysisData} or \code{Analysis} containing analysis results}

\item{cls}{information column to use for sample classes}

\item{rf}{list of additional parameters to pass to \code{randomForest}}

\item{label}{information column to use for sample labels. Set to \code{NULL} for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points.
Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\%
confidence ellipses for each class}

\item{ROC}{should receiver-operator characteristics be plotted?}

\item{seed}{random number seed}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument
of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{pre-treated} data to plot}
}
\description{
A multidimensional scaling (MDS) plot of supervised random forest analysis
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Supervised random forest MDS plot
plotSupervisedRF(d,cls = 'day')
}
