% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSupervisedRF-method.R
\docType{methods}
\name{plotSupervisedRF}
\alias{plotSupervisedRF}
\alias{plotSupervisedRF,Analysis-method}
\title{plotSupervisedRF}
\usage{
plotSupervisedRF(analysis, cls = "class", ...)

\S4method{plotSupervisedRF}{Analysis}(analysis, cls = "class", ...)
}
\description{
plotSupervisedRF
}
