% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSupervisedRF-method.R
\docType{methods}
\name{plotSupervisedRF}
\alias{plotSupervisedRF}
\alias{plotSupervisedRF,AnalysisData-method}
\title{plotSupervisedRF}
\usage{
plotSupervisedRF(x, cls = "class", rf = list(), label = NULL,
  ellipses = T, ROC = T, seed = 1234,
  title = "MDS plot of a supervised random forest",
  legendPosition = "bottom", labelSize = 2)

\S4method{plotSupervisedRF}{AnalysisData}(x, cls = "class",
  rf = list(), label = NULL, ellipses = T, ROC = T, seed = 1234,
  title = "", legendPosition = "bottom", labelSize = 2)
}
\arguments{
\item{x}{object of class Analysis or AnalysisData containing analysis results}

\item{cls}{info column to use for sample classes}

\item{rf}{list of additional parameters to pass to randomForest}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{ellipses}{should multivariate normal distribution 95\% confidence ellipses be plotted for each class?}

\item{ROC}{should reciever-operator characteristics be plotted?}

\item{seed}{random number seed}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
plotSupervisedRF
}
\examples{
library(metaboData)
data(abr1)
p <- analysisParameters('preTreat')
p@preTreat <- list(
  occupancyFilter = list(maximum = list()),
  transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)  

plotSupervisedRF(analysis,label = 'name')
}
