% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotTIC-method.R
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,AnalysisData-method}
\alias{plotTIC,Analysis-method}
\title{plotTIC}
\usage{
plotTIC(analysis, by = "injOrder", colour = "block", ...)

\S4method{plotTIC}{AnalysisData}(analysis, by = "injOrder", colour = "block")

\S4method{plotTIC}{Analysis}(analysis, by = "injOrder", colour = "block", type = "raw")
}
\arguments{
\item{analysis}{object of class Analysis or AnalysisData containing 
analysis results}

\item{by}{info column to plot against}

\item{colour}{info column to provide colour labels}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{preTreated} sample information}
}
\description{
Plot total ion counts of sample raw data.
}
\examples{
\dontrun{
library(metaboData)
data(abr1)
p <- analysisParameters(c('preTreat'))
p@preTreat <- list(
    occupancyFilter = list(maximum = list()),
    transform = list(TICnorm = list())
)
analysis <- metabolyse(abr1$neg,abr1$fact,p)
plotTIC(analysis,by = 'injorder',colour = 'day')
}
}
