% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIC.R
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,AnalysisData-method}
\alias{plotTIC,Analysis-method}
\title{Plot sample total ion counts}
\usage{
plotTIC(analysis, by = "injOrder", colour = "block", ...)

\S4method{plotTIC}{AnalysisData}(analysis, by = "injOrder", colour = "block")

\S4method{plotTIC}{Analysis}(
  analysis,
  by = "injOrder",
  colour = "block",
  type = c("pre-treated", "raw")
)
}
\arguments{
\item{analysis}{S4 object of class \code{AnalysisData} or \code{Analysis}}

\item{by}{information column to plot against}

\item{colour}{information column to provide colour labels}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{pre-treated} sample data}
}
\description{
Plot total ion counts of sample data.
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg,abr1$fact)

## Plot sample TIVs
plotTIC(d,by = 'injorder',colour = 'day')

plotTIC(d,by = 'day',colour = 'day')
}
