% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotUnsupervisedRF-method.R
\docType{methods}
\name{plotUnsupervisedRF}
\alias{plotUnsupervisedRF}
\alias{plotUnsupervisedRF,Analysis-method}
\title{plotUnsupervisedRF}
\usage{
plotUnsupervisedRF(analysis, cls = "class", ...)

\S4method{plotUnsupervisedRF}{Analysis}(analysis, cls = "class",
  seed = 1234, ...)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{...}{additional parameters to pass to randomForest}

\item{seed}{random number seed}
}
\description{
plotUnsupervisedRF
}
