% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnsupervisedRF.R
\name{plotUnsupervisedRF}
\alias{plotUnsupervisedRF}
\alias{plotUnsupervisedRF,AnalysisData-method}
\alias{plotUnsupervisedRF,Analysis-method}
\title{Unsupervised random forest MDS plot}
\usage{
plotUnsupervisedRF(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2,
  ...
)

\S4method{plotUnsupervisedRF}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotUnsupervisedRF}{Analysis}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2,
  type = c("pre-treated", "raw")
)
}
\arguments{
\item{x}{object of class \code{AnalysisData} or \code{Analysis}}

\item{cls}{sample information column to use for sample labelling}

\item{rf}{list of additional parameters to pass to \code{randomForest}}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points.
Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\%
confidence ellipses for each class}

\item{seed}{random number seed}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument
of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}

\item{...}{arguments to pass to the appropriate method}

\item{type}{\code{raw} or \code{pre-treated} data to plot}
}
\description{
A multidimensional scaling (MDS) plot of unsupervised random forest analysis
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Unsupervised random forest MDS plot
plotUnsupervisedRF(d,cls = 'day')
}
