% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnsupervisedRF.R
\docType{methods}
\name{plotUnsupervisedRF,Analysis-method}
\alias{plotUnsupervisedRF,Analysis-method}
\title{plotUnsupervisedRF}
\usage{
\S4method{plotUnsupervisedRF}{Analysis}(analysis, cls = "class",
  seed = 1234, ...)
}
\arguments{
\item{analysis}{object of class Analysis containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{seed}{random number seed}

\item{...}{additional parameters to pass to randomForest}
}
\description{
plotUnsupervisedRF
}
