% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotUnsupervisedRF-method.R
\name{plotUnsupervisedRF}
\alias{plotUnsupervisedRF}
\alias{plotUnsupervisedRF,AnalysisData-method}
\alias{plotUnsupervisedRF,Analysis-method}
\title{plotUnsupervisedRF}
\usage{
plotUnsupervisedRF(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotUnsupervisedRF}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)

\S4method{plotUnsupervisedRF}{Analysis}(
  x,
  cls = "class",
  rf = list(),
  label = NULL,
  shape = FALSE,
  ellipses = TRUE,
  seed = 1234,
  title = "",
  legendPosition = "bottom",
  labelSize = 2
)
}
\arguments{
\item{x}{object of class Analysis containing analysis results}

\item{cls}{info column to use for sample labelling}

\item{rf}{list of additional parameters to pass to randomForest}

\item{label}{info column to use for sample labels. Set to NULL for no labels.}

\item{shape}{TRUE/FALSE use shape aesthetic for plot points. Defaults to TRUE when the number of classes is greater than 12}

\item{ellipses}{TRUE/FALSE, plot multivariate normal distribution 95\% confidence ellipses for each class}

\item{seed}{random number seed}

\item{title}{plot title}

\item{legendPosition}{legend position to pass to legend.position argument of \code{ggplot2::theme}. Set to "none" to remove legend.}

\item{labelSize}{label size. Ignored if \code{label} is \code{NULL}}
}
\description{
plotUnsupervisedRF
}
