% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnsupervisedRF.R
\docType{methods}
\name{plotUnsupervisedRF,Analysis-method}
\alias{plotUnsupervisedRF,Analysis-method}
\title{plotUnsupervisedRF}
\usage{
\S4method{plotUnsupervisedRF}{Analysis}(analysis, cls = "class",
  seed = 1234, ...)
}
\description{
plotUnsupervisedRF
}
