% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisResults.R
\name{preTreated}
\alias{preTreated}
\alias{preTreated<-}
\alias{preTreated,Analysis-method}
\alias{preTreated<-,Analysis-method}
\title{preTreated}
\usage{
preTreated(x)

preTreated(x) <- value

\S4method{preTreated}{Analysis}(x)

\S4method{preTreated}{Analysis}(x) <- value
}
\arguments{
\item{x}{S4 object of class Analysis}

\item{value}{S4 object of class AnalysisData}
}
\description{
Get or set an AnalysisData object from the pre-treated slot of the Analysis class.
}
