% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/preTreatedInfo-method.R
\name{preTreatedInfo}
\alias{preTreatedInfo}
\alias{preTreatedInfo,Analysis-method}
\title{preTreatedInfo-Analysis}
\usage{
preTreatedInfo(x)

\S4method{preTreatedInfo}{Analysis}(x)
}
\arguments{
\item{x}{Analysis object}
}
\description{
Extract pre-treated info from an Analysis object
}
