% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTreatMethods.R
\name{preTreatmentMethods}
\alias{preTreatmentMethods}
\title{preTreatmentMethods}
\usage{
preTreatmentMethods(element)
}
\arguments{
\item{element}{pre-treatment element}
}
\description{
Return names of available methods for a given pre-treatment 
element.
}
