% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/QCMethods.R
\docType{methods}
\name{QCrsdFilter}
\alias{QCrsdFilter}
\alias{QCrsdFilter,AnalysisData-method}
\title{QCrsdFilter}
\usage{
QCrsdFilter(d, cls = "class", QCidx = "QC", RSDthresh = 0.5)

\S4method{QCrsdFilter}{AnalysisData}(d, cls = "class", QCidx = "QC",
  RSDthresh = 0.5)
}
\arguments{
\item{d}{S4 object of class AnalysisData}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{RSDthresh}{RSD threshold for filtering}
}
\description{
QC relative standard deviation (RSD) filtering..
}
