% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{randomForest}
\alias{randomForest}
\alias{randomForest,AnalysisData-method}
\title{Random forest}
\usage{
randomForest(
  x,
  cls = "class",
  rf = list(),
  reps = 1,
  binary = FALSE,
  comparisons = list(),
  perm = 0,
  returnModels = FALSE,
  seed = 1234
)

\S4method{randomForest}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  reps = 1,
  binary = FALSE,
  comparisons = list(),
  perm = 0,
  returnModels = FALSE,
  seed = 1234
)
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{vector of sample information columns to use for response variable information. Set to NULL for unsupervised.}

\item{rf}{named list of arguments to pass to \code{randomForest::randomForest}}

\item{reps}{number of repetitions to perform}

\item{binary}{TRUE/FALSE should binary comparisons be performed. Ignored for unsupervised and regression. Ignored if \code{comparisons} specified.}

\item{comparisons}{list of comparisons to perform.
Ignored for unsupervised and regression. See details.}

\item{perm}{number of permutations to perform. Ignored for unsupervised.}

\item{returnModels}{TRUE/FALSE should model objects be returned.}

\item{seed}{random number seed}
}
\value{
An S4 object of class \code{RandomForest}.
}
\description{
Perform random forest on an \code{AnalysisData} object
}
\details{
Specified class comparisons should be given as a list named
according to \code{cls}. Comparisons should be given as class names
separated by '~' (eg. '1~2~H').
}
\examples{
library(metaboData)

x <- analysisData(abr1$neg[,200:300],abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
       
rf <- randomForest(x,cls = 'day')

plotMDS(rf,cls = 'day')
}
