% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/randomForest.R
\name{randomForest}
\alias{randomForest}
\alias{randomForest,AnalysisData-method}
\title{randomForest}
\usage{
randomForest(
  x,
  cls = "class",
  rf = list(),
  reps = 1,
  binary = FALSE,
  comparisons = list(),
  perm = 0,
  returnModels = FALSE,
  seed = 1234,
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)

\S4method{randomForest}{AnalysisData}(
  x,
  cls = "class",
  rf = list(),
  reps = 1,
  binary = F,
  comparisons = list(),
  perm = 0,
  returnModels = F,
  seed = 1234,
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{vector of info columns to use for Response information. Set to NULL for unsupervised.}

\item{rf}{named list of arguments to pass to randomForest::randomForest}

\item{reps}{number of repetitions to perform}

\item{binary}{TRUE/FALSE should bnary comparisons be performed. Ignored for unsupervised and regression. Ignored if \code{comparisons} specified.}

\item{comparisons}{list of comparisons to perform. Ignored for unsupervised and regression. See details.}

\item{perm}{number of permutations to perform. Ignored for unsupervised.}

\item{returnModels}{TRUE/FALSE should model objects be returned.}

\item{seed}{random number seed}

\item{nCores}{number of cores to use for parallisation.}

\item{clusterType}{cluster type for parallisation}
}
\description{
Perform random forest on an AnalysisData object
}
\details{
Specified class comparisons should be given as a list named according to \code{cls}. Comparisons should be given as class names separated by '~' (eg. '1~2~H').
}
\examples{
\dontrun{
library(metaboData)
library(magrittr)

data(abr1)
x <- analysisData(abr1$neg,abr1$fact) \%>\%
       occupancyMaximum(cls = 'day') \%>\%
       transformTICnorm()
rf <- randomForest(x,cls = 'day')
}
}
