% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/rawData-method.R
\docType{methods}
\name{raw}
\alias{raw}
\alias{raw,Analysis-method}
\alias{raw<-}
\title{raw}
\usage{
raw(x)

\S4method{raw}{Analysis}(x)

raw(x) <- value
}
\arguments{
\item{x}{S4 object of class Analysis}

\item{value}{S4 object of class AnalysisData}
}
\description{
Get or set an AnalysisData object from the rawData slot of the Analysis class.
}
