% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reAnalyse.R
\name{reAnalyse}
\alias{reAnalyse}
\title{reAnalyse}
\usage{
reAnalyse(analysis, parameters = analysisParameters(), verbose = TRUE)
}
\arguments{
\item{analysis}{an object of class Analysis containing previous 
analysis results}

\item{parameters}{an object of class Parameters containing parameters for 
re-analysis}

\item{verbose}{should output be printed to the console}
}
\description{
Re-analyse an object of class Analysis using 
specified parameters.
}
\examples{
library(metaboData)

## Generate analysis parameters
p <- analysisParameters(c('pre-treatment','modelling'))

## Alter pre-treatment and modelling parameters to use different methods
parameters(p,'pre-treatment') <- preTreatmentParameters(
  list(occupancyFilter = 'maximum',
       transform = 'TICnorm')
)
parameters(p,'modelling') <- modellingParameters('anova')

## Change "cls" and "nCores" parameters 
changeParameter(p,'cls') <- 'day'
changeParameter(p,'nCores') <- 2

## Run analysis using a subset of the abr1 negative mode data set
analysis <- metabolyse(abr1$neg[,1:200],
                       abr1$fact,
                       p)
                       
## Re-analyse to include correlation analysis
analysis <- reAnalyse(analysis,
                      parameters = analysisParameters('correlations'))
}
