% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reAnalyse.R
\name{reAnalyse}
\alias{reAnalyse}
\title{reAnalyse}
\usage{
reAnalyse(analysis, parameters = analysisParameters())
}
\arguments{
\item{analysis}{an object of class Analysis containing previous analysis results}

\item{parameters}{an object of class Parameters containing parameters for re-analysis}
}
\description{
Re-analyse an object of class Analysis using specified parameters.
}
\seealso{
\link{metabolyse} \link{analysisParameters} \linkS4class{Parameters} \linkS4class{Analysis}
}
