% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/QCMethods.R
\name{rsd}
\alias{rsd}
\alias{rsd,AnalysisData-method}
\title{rsd}
\usage{
rsd(x, cls = "class")

\S4method{rsd}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{info column to use for class structure}
}
\description{
Calculate relative standard deviation values for each 
feature per class for a given info column.
}
