% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,AnalysisPlot-method}
\alias{show,AnalysisPlot-method}
\title{show-AnalysisPlot}
\usage{
\S4method{show}{AnalysisPlot}(object)
}
\arguments{
\item{object}{S4 object of class AnalysisData}
}
\description{
show method for AnalysisData class.
}
