% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/analysisResults.R
\name{sinfo}
\alias{sinfo}
\alias{sinfo<-}
\alias{sinfo,AnalysisData-method}
\alias{sinfo<-,AnalysisData-method}
\alias{sinfo,Analysis-method}
\alias{sinfo<-,Analysis-method}
\title{sinfo}
\usage{
sinfo(x, ...)

sinfo(x, ...) <- value

\S4method{sinfo}{AnalysisData}(x)

\S4method{sinfo}{AnalysisData}(x) <- value

\S4method{sinfo}{Analysis}(x, type = "raw", value)

\S4method{sinfo}{Analysis}(x, type = "raw") <- value
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{value}{tibble containing sample info}

\item{type}{Sample information type to extract or set. Should be one of "raw" or "pre-treated"}
}
\description{
Return sample info from an AnalysisData object.
}
