% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\alias{split,AnalysisData-method}
\title{Split an \code{AnalysisData} object}
\usage{
split(x, cls = "class")

\S4method{split}{AnalysisData}(x, cls = "class")
}
\arguments{
\item{x}{S4 object of class \code{AnalysisData}}

\item{cls}{sample information column to use for splitting}
}
\value{
A list of \code{AnalysisData} objects.
}
\description{
Split an object of class \code{AnalysisData} into a list based
a class grouping variable.
}
\examples{
library(metaboData)

d <- analysisData(abr1$neg,abr1$fact)

## Split the data set based on the 'day' class information column
d <- split(d,cls = 'day')

print(d)
}
