% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transform.R
\name{transformCenter}
\alias{transformCenter}
\alias{transformAuto}
\alias{transformRange}
\alias{transformPareto}
\alias{transformLevel}
\alias{transformVast}
\alias{transformLn}
\alias{transformLog10}
\alias{transformSQRT}
\alias{transformArcSine}
\alias{transformTICnorm}
\alias{transformArcSine,AnalysisData-method}
\alias{transformAuto,AnalysisData-method}
\alias{transformCenter,AnalysisData-method}
\alias{transformLevel,AnalysisData-method}
\alias{transformLn,AnalysisData-method}
\alias{transformLog10,AnalysisData-method}
\alias{transformPareto,AnalysisData-method}
\alias{transformRange,AnalysisData-method}
\alias{transformSQRT,AnalysisData-method}
\alias{transformTICnorm,AnalysisData-method}
\alias{transformVast,AnalysisData-method}
\title{Scaling, transformation and normalisation methods}
\usage{
transformCenter(d)

transformAuto(d)

transformRange(d)

transformPareto(d)

transformLevel(d)

transformVast(d)

transformLn(d, add = 1)

transformLog10(d, add = 1)

transformSQRT(d)

transformArcSine(d)

transformTICnorm(d)

\S4method{transformArcSine}{AnalysisData}(d)

\S4method{transformAuto}{AnalysisData}(d)

\S4method{transformCenter}{AnalysisData}(d)

\S4method{transformLevel}{AnalysisData}(d)

\S4method{transformLn}{AnalysisData}(d, add = 1)

\S4method{transformLog10}{AnalysisData}(d, add = 1)

\S4method{transformPareto}{AnalysisData}(d)

\S4method{transformRange}{AnalysisData}(d)

\S4method{transformSQRT}{AnalysisData}(d)

\S4method{transformTICnorm}{AnalysisData}(d)

\S4method{transformVast}{AnalysisData}(d)
}
\arguments{
\item{d}{S4 object of class AnalysisData}
}
\description{
Methods for data pre-treatment scaling, transformation and normalisation.
}
\section{Methods}{

\itemize{
\item \code{transformArcSine}: Arc-Sine transformation.
\item \code{transformAuto}: Auto scaling.
\item \code{transformCenter}: Mean centring.
\item \code{transformLevel}: Level scaling.
\item \code{transformLn}: Natural logarithmic transformation.
\item \code{transformLog10}: Logarithmic transformation.
\item \code{transformPareto}: Pareto scaling.
\item \code{transformRange}: Range scaling. Also known as min-max scaling.
\item \code{transformSQRT}: Square root transformation.
\item \code{transformTICnorm}: Total ion count normalisation.
\item \code{transformVast}: Vast scaling.
}
}

\concept{Pre-treatment}
