% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\name{transformArcSine}
\alias{transformArcSine}
\alias{transformArcSine,AnalysisData-method}
\title{transformArcSine}
\usage{
transformArcSine(d)

\S4method{transformArcSine}{AnalysisData}(d)
}
\arguments{
\item{d}{S4 object of class AnalysisData}
}
\description{
Arc-sine transformation of sample data.
}
