% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\name{transformCenter}
\alias{transformCenter}
\alias{transformCenter,AnalysisData-method}
\title{transformCenter}
\usage{
transformCenter(d)

\S4method{transformCenter}{AnalysisData}(d)
}
\arguments{
\item{d}{S4 object of class AnalysisData}
}
\description{
Mean center sample data.
}
