% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformLn}
\alias{transformLn}
\alias{transformLn,AnalysisData-method}
\title{transformLn}
\usage{
transformLn(dat, add = 1)

\S4method{transformLn}{AnalysisData}(dat, add = 1)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Natural logarithmic transformation of sample data.
}
