% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformLog10}
\alias{transformLog10}
\alias{transformLog10,AnalysisData-method}
\title{transformLog10}
\usage{
transformLog10(dat, add = 1)

\S4method{transformLog10}{AnalysisData}(dat, add = 1)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Logarithmic transformation of sample data.
}
