% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\docType{methods}
\name{transformTICnorm}
\alias{transformTICnorm}
\alias{transformTICnorm,AnalysisData-method}
\title{transformTICnorm}
\usage{
transformTICnorm(dat)

\S4method{transformTICnorm}{AnalysisData}(dat)
}
\arguments{
\item{dat}{S4 object of class AnalysisData}
}
\description{
Total ion count normalisation of sample data.
}
