% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transformMethods.R
\name{transformVast}
\alias{transformVast}
\alias{transformVast,AnalysisData-method}
\title{transformVast}
\usage{
transformVast(d)

\S4method{transformVast}{AnalysisData}(d)
}
\arguments{
\item{d}{S4 object of class AnalysisData}
}
\description{
Vast scaling of sample data.
}
