% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/univariateMethods.R
\docType{methods}
\name{ttest}
\alias{ttest}
\alias{ttest,AnalysisData-method}
\title{ttest}
\usage{
ttest(x, cls, pAdjust = "bonferroni", returnModels = F,
  nCores = detectCores() * 0.75, clusterType = clusterType(), ...)

\S4method{ttest}{AnalysisData}(x, cls, pAdjust = "bonferroni",
  returnModels = F, nCores = detectCores() * 0.75,
  clusterType = getClusterType(), ...)
}
\description{
ttest
}
