% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/univariateMethods.R
\name{ttest}
\alias{ttest}
\alias{ttest,AnalysisData-method}
\title{ttest}
\usage{
ttest(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE,
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)

\S4method{ttest}{AnalysisData}(
  x,
  cls = "class",
  pAdjust = "bonferroni",
  comparisons = list(),
  returnModels = FALSE,
  nCores = detectCores() * 0.75,
  clusterType = getClusterType()
)
}
\arguments{
\item{x}{S4 object of class AnalysisData}

\item{cls}{vector of sample info column names to analyse}

\item{pAdjust}{p value adjustment method}

\item{comparisons}{named list of binary comparisons to analyse}

\item{returnModels}{should models be returned}

\item{nCores}{number of cores to use for parallelisation}

\item{clusterType}{cluster type to use for parallelisation 
(`?parallel::makeCluster`)}
}
\description{
Welch t-test
}
